update:removed bloat and unused js
