+++ src/app/account/(tabs)/events/e/[id]/components/event-photo.tsx
@@ -1,4 +1,4 @@
-import { Iconx } from "@/icons/icon";
+import { Iconx } from "@/icons";
 import { Button } from "@nextui-org/react";
 
 export const EventPhoto = () => (
