+++ src/app/account/(tabs)/events/event-card.tsx
@@ -3,7 +3,7 @@
 import { getUserID } from "@/app/actions";
 import { type XEvent } from "@/app/types";
 import { useMoment } from "@/hooks/useMoment";
-import { Iconx } from "@/icons/icon";
+import { Iconx } from "@/icons";
 import { cn } from "@/lib/utils";
 import { BtnIcon } from "@/ui/button";
 import { opts } from "@/utils/helpers";
