+++ src/app/account/(tabs)/events/l/[id]/content.tsx
@@ -5,7 +5,7 @@ import { use, useCallback, useEffect } from "react";
 import { Image } from "@nextui-org/react";
 import { ScanCode } from "./components/scanner";
 import { cn } from "@/lib/utils";
-import { Iconx } from "@/icons/icon";
+import { Iconx } from "@/icons";
 
 export const Content = () => (
   <LiveViewCtxProvider>
