+++ src/components/ui/sidebar.tsx
@@ -32,7 +32,7 @@ import {
   useMemo,
   useState,
 } from "react";
-import { Iconx } from "@/icons/icon";
+import { Iconx } from "@/icons";
 
 const SIDEBAR_COOKIE_NAME = "sidebar_state";
 const SIDEBAR_COOKIE_MAX_AGE = 60 * 60 * 24 * 7;
