+++ src/icons/index.tsx
@@ -1,19 +1,31 @@
-// import type { SVGProps } from "react";
-// import { icons } from "./list";
+import type { FC } from "react";
+import { icons } from "./icons";
+import type { IconProps } from "./types";
 
-// export type IconName = keyof typeof icons;
-// export type IconProps = SVGProps<SVGSVGElement> & { name: IconName };
+export const Iconx: FC<Omit<IconProps, "content">> = ({
+  name,
+  className,
+  size = 24,
+  color = "currentColor",
+  solid = false,
+  ...props
+}) => {
+  const icon = icons[name];
 
-// export function Icon({
-//   name,
-//   fill = "none",
-//   stroke = "currentColor",
-//   ...props
-// }: IconProps) {
-//   const Component = icons[name];
-//   if (!Component) {
-//     console.error(`Icon "${name}" not found.`);
-//     return null;
-//   }
-//   return <Component fill={fill} focusable={false} stroke={stroke} {...props} />;
-// }
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox={"0 0 24 24"}
+      width={size}
+      height={size}
+      className={className}
+      fill={solid ? color : "none"}
+      stroke={color}
+      strokeWidth="1"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      {...props}
+      dangerouslySetInnerHTML={{ __html: icon.symbol }}
+    />
+  );
+};
