--- src/icons/list.tsx
@@ -1,2636 +1,0 @@
-// import type { SVGProps } from "react";
-
-// function Activity(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1"
-//         color="currentColor"
-//       >
-//         <path d="M4.318 19.682C3 18.364 3 16.242 3 12s0-6.364 1.318-7.682S7.758 3 12 3s6.364 0 7.682 1.318S21 7.758 21 12s0 6.364-1.318 7.682S16.242 21 12 21s-6.364 0-7.682-1.318"></path>
-//         <path d="M6 12h2.5l2-4l3 8l2-4H18"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function AddUsers(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M12 7.5a3.5 3.5 0 1 1-7 0a3.5 3.5 0 0 1 7 0m1.5 3.5a3.5 3.5 0 1 0 0-7m-.357 16H3.857C2.831 20 2 19.233 2 18.286C2 15.919 4.079 14 6.643 14h3.714a4.9 4.9 0 0 1 2.786.857M19 14v6m3-3h-6"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function Airplane(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 64 64"
-//       {...props}
-//     >
-//       <path
-//         fill={props.fill}
-//         d="M61.944 5.971c-.375-1.864-2.486-3.106-5.947-3.497C53.7 2.215 51.649 2 49.833 2c-5.145 0-8.582 1.662-18.701 7.035c-1.787.056-3.555.266-4.471.566c-1.021.33-.924 1.055.079 1.769a1254 1254 0 0 1-8.094 4.264c-3.643-1.093-10.278-3.099-12.993-4.016C3.236 10.804 2 13.49 2 13.49s2.46 6.035 3.937 9.091c-.331.378-.596.801-.717 1.28c-.22.868-.137 2.317.818 3.544c1.125 1.444 3.082 2.177 5.819 2.177c2.972 0 6.967-.87 11.933-2.576l-1.033 1.554c-1.571.997-3.236 2.319-5.023 4.119c-.313.316-.819.947-.695 1.636c.054.3.255.719.969 1.042c.063.018.134.03.2.046l-.061.092c-1.287 2.074-.387 3.132 2.001 2.353c1.447-.474 3.209-1.461 4.727-2.58c2.551-.463 5.911-1.394 10.234-3.167c1.045-.429 1.543-2.24 1.111-4.036c-.583-2.453-1.668-2.825-2.266-2.826l3.141-3.715c6.486-3.015 13.93-6.859 22.34-11.552c2.591-1.447 2.685-3.128 2.509-4.001M18.741 34.599a5 5 0 0 1-.42-.093c-.161-.073-.359-.195-.388-.354c-.035-.194.144-.528.446-.832a32 32 0 0 1 2.991-2.675zm16.594-6.325c.363 1.504-.104 2.792-.571 2.983c-3.374 1.385-6.14 2.244-8.382 2.768c.365-.336.707-.673.984-.999l5.744-6.792l.171-.02l.667-.068c.553 0 1.071.795 1.387 2.128m23.213-19.89c-7.144 3.985-13.297 7.196-18.606 9.774l2.469-2.919c1.583-1.87.997-2.564-1.302-1.538l-8.874 3.958c-2.3 1.026-3.362 1.565-4.649 3.638l-2.112 3.179c-6.595 2.459-10.863 3.288-13.616 3.288c-5.228 0-4.993-2.988-4.875-3.455l14.068-7.135c.474-.241.516-.6.094-.793h-.001c4.06-2.128 7.412-3.917 10.233-5.417a830 830 0 0 1 3.461-1.83c7.807-4.098 10.76-5.316 14.996-5.316c1.664 0 3.526.188 5.961.463c4.28.482 5.877 2.361 2.753 4.103"
-//       ></path>
-//       <path
-//         fill={props.fill}
-//         d="M33.456 26.517c-.164.591-.204 1.34-.082 2.122c.186 1.172.688 2.089 1.254 2.437c.164-.595.204-1.344.078-2.127c-.184-1.173-.684-2.087-1.25-2.432m15.06-19.236s.492.777 1.864.389c1.188-.336 2.363-1.488 3.031-2.751l-1.953-.435c-.41.686-2.942 2.797-2.942 2.797"
-//       ></path>
-//       <circle cx="21.351" cy="19.661" r=".864" fill={props.fill}></circle>
-//       <circle cx="24.646" cy="17.961" r=".865" fill={props.fill}></circle>
-//       <circle cx="27.942" cy="16.262" r=".865" fill={props.fill}></circle>
-//       <circle cx="31.238" cy="14.563" r=".865" fill={props.fill}></circle>
-//       <circle cx="34.533" cy="12.862" r=".864" fill={props.fill}></circle>
-//       <path
-//         fill={props.fill}
-//         d="M37.433 10.396a.864.864 0 1 0 .793 1.535a.86.86 0 0 0 .372-1.164a.86.86 0 0 0-1.165-.371"
-//       ></path>
-//       <circle cx="41.124" cy="9.463" r=".865" fill={props.fill}></circle>
-//     </svg>
-//   );
-// }
-
-// function Airplane2(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 70 70"
-//       {...props}
-//     >
-//       <path d="M24 39.5L8.119 34.508l-4.437 1.375L13.3 45" fill="#E62E2F" />
-
-//       <path
-//         fill="#d0cfce"
-//         d="m42.589 30.5l13.635-4.637s13.5-.229 10.33 3.255S41.127 41.314 34.551 43.77s-14.712 3.75-20.337 3.86c-4.924.095.59-2.892.59-2.892s6.255-3.55 9.024-4.757l2.635-1.949"
-//       ></path>
-//       <path
-//         fill="#9b9b9a"
-//         d="m48.2 35l-25.755-2.359l-2.093 1.493L38 40.5"
-//       ></path>
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeMiterlimit="10"
-//         strokeWidth="1"
-//       >
-//         <path d="M19.974 39.139L7.542 34.736l-3.067 1.073l8.169 7.322m34.407-8.438l-24.606-2.052l-2.093 1.493l16.754 5.979"></path>
-//         <path d="m43.174 31.323l11.15-4.486a10.3 10.3 0 0 1 4.172-.74c3.7.113 10.39.667 8.057 3.23c-3.17 3.484-25.425 12.196-32.001 14.651s-14.712 3.75-20.337 3.86c-4.924.094.59-2.892.59-2.892s6.255-3.55 9.024-4.757l2.784-1.213"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function AlertTriLine(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="currentColor"
-//         d="m21.171 15.398l-5.912-9.854C14.483 4.251 13.296 3.511 12 3.511s-2.483.74-3.259 2.031l-5.912 9.856c-.786 1.309-.872 2.705-.235 3.83C3.23 20.354 4.472 21 6 21h12c1.528 0 2.77-.646 3.406-1.771s.551-2.521-.235-3.831M12 17.549c-.854 0-1.55-.695-1.55-1.549c0-.855.695-1.551 1.55-1.551s1.55.696 1.55 1.551c0 .854-.696 1.549-1.55 1.549m1.633-7.424c-.011.031-1.401 3.468-1.401 3.468c-.038.094-.13.156-.231.156s-.193-.062-.231-.156l-1.391-3.438a1.8 1.8 0 0 1-.129-.655c0-.965.785-1.75 1.75-1.75a1.752 1.752 0 0 1 1.633 2.375"
-//       ></path>
-//     </svg>
-//   );
-// }
-
-// function ArrowRight(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none">
-//         <path
-//           stroke="none"
-//           fill="currentColor"
-//           d="m15.06 5.283l5.657 5.657a1.5 1.5 0 0 1 0 2.12l-5.656 5.658a1.5 1.5 0 0 1-2.122-2.122l3.096-3.096H4.5a1.5 1.5 0 0 1 0-3h11.535L12.94 7.404a1.5 1.5 0 0 1 2.122-2.121Z"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function ArrowRightUp(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none" stroke="currentColor" strokeWidth="1.5">
-//         <path
-//           strokeLinecap="round"
-//           strokeLinejoin="round"
-//           d="m9 15l6-6m0 0h-4.5M15 9v4.5"
-//         ></path>
-//         <path d="M2 12c0-4.714 0-7.071 1.464-8.536C4.93 2 7.286 2 12 2s7.071 0 8.535 1.464C22 4.93 22 7.286 22 12s0 7.071-1.465 8.535C19.072 22 16.714 22 12 22s-7.071 0-8.536-1.465C2 19.072 2 16.714 2 12Z"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function ArrowRightDouble(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="currentColor" fillRule="evenodd" clipRule="evenodd">
-//         <path d="M6.41 7.47a.75.75 0 0 1 1.06 0L12 12l-4.53 4.53a.75.75 0 0 1-1.06-1.06L9.878 12l-3.47-3.47a.75.75 0 0 1 0-1.06"></path>
-//         <path d="M12.22 7.47a.75.75 0 0 1 1.06 0L17.81 12l-4.53 4.53a.75.75 0 1 1-1.06-1.06L15.69 12l-3.47-3.47a.75.75 0 0 1 0-1.06"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function ArrowReturnUp(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         color="currentColor"
-//       >
-//         <path d="M2.5 12c0-4.478 0-6.718 1.391-8.109S7.521 2.5 12 2.5c4.478 0 6.718 0 8.109 1.391S21.5 7.521 21.5 12c0 4.478 0 6.718-1.391 8.109S16.479 21.5 12 21.5c-4.478 0-6.718 0-8.109-1.391S2.5 16.479 2.5 12"></path>
-//         <path d="M14.175 11.352c.672-.79 2.038-1.769 1.795-2.233m0 0c.135-.474-1.187-1.41-1.798-2.124m1.798 2.124c-2.755-.057-4.441.259-6.068 1.794c-1.695 1.599-2.01 3.324-1.873 6.092"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function ArrowVertical(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1"
-//         d="M9 5c.59-.607 2.16-3 3-3s2.41 2.393 3 3M9 19c.59.607 2.16 3 3 3s2.41-2.393 3-3M12 2.231V21.77"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function Bell(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none">
-//         <path
-//           fill="currentColor"
-//           d="M12 2a7 7 0 0 0-7 7v3.528a1 1 0 0 1-.105.447l-1.717 3.433A1.1 1.1 0 0 0 4.162 18h15.676a1.1 1.1 0 0 0 .984-1.592l-1.716-3.433a1 1 0 0 1-.106-.447V9a7 7 0 0 0-7-7m0 19a3 3 0 0 1-2.83-2h5.66A3 3 0 0 1 12 21"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function Bookmark(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//       >
-//         <path d="M4.421 8.45v11.884c0 .763.88 1.19 1.48.718l4.883-3.843a.914.914 0 0 1 1.13 0l4.884 3.843a.914.914 0 0 0 1.48-.718V8.45a3.2 3.2 0 0 0-3.2-3.2H7.621a3.2 3.2 0 0 0-3.2 3.2Z"></path>
-//         <path
-//           strokeLinecap="round"
-//           d="M21.278 16.334V8.25a6 6 0 0 0-6-6H8.421"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function Bookmark2(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//       >
-//         <path d="M20.5 16.929V10c0-3.771 0-5.657-1.172-6.828S16.271 2 12.5 2h-1C7.729 2 5.843 2 4.672 3.172S3.5 6.229 3.5 10v9.5"></path>
-//         <path
-//           fill="currentColor"
-//           d="M8 3v3.69c0 .74 0 1.11.238 1.254q.014.01.032.017c.248.126.571-.064 1.219-.445c.484-.284.725-.427.992-.432h.037c.268.005.51.148.993.432c.648.381.971.571 1.22.445l.03-.017C13 7.8 13 7.43 13 6.69V3m7.5 14H6a2.5 2.5 0 0 0 0 5h14.5"
-//         ></path>
-//         <path d="M20.5 17a2.5 2.5 0 0 0 0 5"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function BookmarkPlus(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M11 2C7.229 2 5.343 2 4.172 3.129C3 4.257 3 6.074 3 9.708v8.273c0 2.306 0 3.459.773 3.871c1.496.8 4.304-1.867 5.637-2.67c.773-.465 1.16-.698 1.59-.698s.817.233 1.59.698c1.333.803 4.14 3.47 5.637 2.67c.773-.412.773-1.565.773-3.871V13m-2-3V2m-4 4h8"
-//       ></path>
-//     </svg>
-//   );
-// }
-
-// function BookmarkCheck(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         color="currentColor"
-//       >
-//         <path d="M4 17.98V9.709c0-3.634 0-5.45 1.172-6.58S8.229 2 12 2s5.657 0 6.828 1.129C20 4.257 20 6.074 20 9.708v8.273c0 2.306 0 3.459-.773 3.871c-1.497.8-4.304-1.867-5.637-2.67c-.773-.465-1.16-.698-1.59-.698s-.817.233-1.59.698c-1.333.803-4.14 3.47-5.637 2.67C4 21.44 4 20.287 4 17.981"></path>
-//         <path
-//           stroke={"#02C7BE"}
-//           strokeWidth="2"
-//           d="M9 11.333s.875 0 1.75 1.667c0 0 2.78-4.167 5.25-5"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function BouncyArrow(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       width="60"
-//       height="40"
-//       viewBox="0 0 202 139"
-//       fill="none"
-//       xmlns="http://www.w3.org/2000/svg"
-//       {...props}
-//     >
-//       <path
-//         fill-rule="evenodd"
-//         clip-rule="evenodd"
-//         d="M79.3953 0.222873C68.6059 1.25504 56.3155 5.75905 46.5582 12.257C35.8158 19.3884 26.3869 29.3816 19.28 41.1812C17.0987 44.7938 12.9003 53.0746 11.1411 57.2033C6.87232 67.2904 3.58861 78.1047 1.31348 89.6931C0.281455 94.9008 0 97.2232 0 100.249C0.023455 104.824 0.797459 109.515 1.57148 109.515C1.94676 109.515 1.89985 110.008 2.01713 104.683C2.11095 99.6864 2.43932 97.0121 3.65898 91.6871C7.52906 74.6328 11.3053 63.5136 17.4036 51.3387C22.681 40.7824 28.2633 32.8769 35.7454 25.4172C47.0742 14.1102 60.6078 6.90851 75.4079 4.28117C78.3398 3.76509 85.3294 3.57742 88.7304 3.9293C100.388 5.14913 111.341 10.3334 116.313 17.0425C118.682 20.2563 120.606 25.253 121.004 29.2643C121.38 32.8065 120.723 37.5921 119.48 40.7355C118.893 42.2368 118.87 43.3159 119.456 44.0666C120.019 44.8172 120.911 45.2395 121.872 45.2395C123.256 45.2395 124.194 44.395 125.133 42.3072C126.071 40.2663 128.111 37.6859 130.879 35.082C136.25 30.0385 143.592 26.5432 152.27 24.9245C154.967 24.4085 162.074 24.2208 165.147 24.5961C170.049 25.1591 176.522 27.1765 180.463 29.3816C184.31 31.5398 189.165 35.8327 191.323 39.023C194.442 43.6443 195.615 48.4767 195.146 54.7167C194.278 66.0471 190.455 76.4626 182.855 88.051C178.633 94.5021 172.488 100.953 166.015 105.786C158.11 111.674 147.509 117.491 139.745 120.189C138.455 120.635 137.306 121.01 137.188 121.01C137.048 121.01 137.704 120.142 138.619 119.063C143.123 113.879 145.257 108.624 143.803 106.278C142.771 104.66 140.519 104.073 138.924 105.058C137.869 105.692 135.57 108.671 133.858 111.603C132.685 113.597 131.981 114.465 130.175 116.178C124.077 121.948 123.256 123.098 123.256 125.819C123.256 127.883 123.913 129.314 125.602 131.027C128.768 134.17 134.749 135.625 150.112 136.962C153.138 137.22 156.398 137.595 157.336 137.806C158.556 138.088 159.142 138.111 159.306 137.947C160.01 137.243 157.172 134.991 152.809 132.81C149.08 130.933 146.688 129.995 142.677 128.869C140.965 128.376 139.557 127.907 139.557 127.836C139.557 127.743 140.636 127.32 141.973 126.898C147.673 125.045 153.888 122.066 160.784 117.843C176.03 108.577 185.67 98.3258 193.363 83.242C196.764 76.5799 198.547 71.6536 199.907 65.0383C200.845 60.6282 201.103 58.0947 201.08 53.5672C201.08 48.9459 200.658 46.1544 199.485 43.0344C197.655 38.1785 194.184 33.393 189.915 29.8273C183.395 24.4085 176.522 21.4058 167.352 19.9983C164.209 19.5291 157.735 19.4587 154.569 19.881C144.366 21.2416 135.312 24.948 128.651 30.5311L126.774 32.1028L126.61 30.1792C126.024 23.3997 122.928 16.5968 118.331 12.0224C116.196 9.91118 112.42 7.28384 109.23 5.66521C100.645 1.34887 89.2698 -0.715462 79.3953 0.222873Z"
-//         fill="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-
-// function BurgerRight(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="currentColor"
-//         d="M4 5.5a.5.5 0 0 0 0 1h12a.5.5 0 0 0 0-1zm1 6a.5.5 0 0 0 0 1h15.793l-1.647 1.646a.5.5 0 0 0 .708.708l2.5-2.5a.5.5 0 0 0 0-.708l-2.5-2.5a.5.5 0 0 0-.708.708l1.647 1.646zM3.5 18a.5.5 0 0 1 .5-.5h12a.5.5 0 0 1 0 1H4a.5.5 0 0 1-.5-.5"
-//       ></path>
-//     </svg>
-//   );
-// }
-
-// function Calendar(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M18 2v2M6 2v2m5.996 9h.008m-.008 4h.008m3.987-4H16m-8 0h.009M8 17h.009M3.5 8h17m-18 4.243c0-4.357 0-6.536 1.252-7.89C5.004 3 7.02 3 11.05 3h1.9c4.03 0 6.046 0 7.298 1.354C21.5 5.707 21.5 7.886 21.5 12.244v.513c0 4.357 0 6.536-1.252 7.89C18.996 22 16.98 22 12.95 22h-1.9c-4.03 0-6.046 0-7.298-1.354C2.5 19.293 2.5 17.114 2.5 12.756zM3 8h18"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function CarDashboard(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="currentColor"
-//         d="M12 15q.11 0 .217.012l3.911-4.347a.5.5 0 1 1 .744.67l-3.664 4.07A2 2 0 0 1 14 17h-4a2 2 0 0 1 2-2m-7 .5a.5.5 0 0 0 0 1h1.5a.5.5 0 0 0 0-1zm14.5.5a.5.5 0 0 0-.5-.5h-1.5a.5.5 0 0 0 0 1H19a.5.5 0 0 0 .5-.5M18 13a1 1 0 1 0 0-2a1 1 0 0 0 0 2M7 12a1 1 0 1 1-2 0a1 1 0 0 1 2 0m1-2.5a1 1 0 1 0 0-2a1 1 0 0 0 0 2m9-1a1 1 0 1 1-2 0a1 1 0 0 1 2 0M12.5 6a.5.5 0 0 0-1 0v1.5a.5.5 0 0 0 1 0z"
-//       ></path>
-//       <path
-//         fill="currentColor"
-//         d="M12 3.5C6.201 3.5 1.5 8.201 1.5 14c0 1.464.3 2.859.842 4.126l.004.009c.212.497.346.81.77 1.261c.165.176.402.36.632.511s.493.298.719.38c.582.213 1.084.213 1.871.213h11.324c.787 0 1.29 0 1.871-.212c.226-.083.488-.229.718-.38c.23-.153.468-.336.633-.512c.424-.45.558-.764.77-1.261l.004-.009c.542-1.267.842-2.662.842-4.126c0-5.799-4.701-10.5-10.5-10.5M2.5 14a9.5 9.5 0 1 1 19 0a9.5 9.5 0 0 1-.761 3.733c-.206.481-.284.66-.584.979c-.09.095-.256.23-.454.36a2.7 2.7 0 0 1-.51.276c-.4.146-.728.152-1.588.152H6.397c-.86 0-1.187-.006-1.587-.152a2.8 2.8 0 0 1-.51-.275a2.8 2.8 0 0 1-.455-.361c-.3-.32-.378-.498-.584-.98A9.5 9.5 0 0 1 2.5 14"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function Chats(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none" fillRule="evenodd">
-//         <path
-//           fill="currentColor"
-//           d="M19 3a3 3 0 0 1 3 3v10a3 3 0 0 1-3 3H7.333L4 21.5c-.824.618-2 .03-2-1V6a3 3 0 0 1 3-3zm-8 9H8a1 1 0 1 0 0 2h3a1 1 0 1 0 0-2m5-4H8a1 1 0 0 0-.117 1.993L8 10h8a1 1 0 0 0 .117-1.993z"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function ChartIcon(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M12 18v4m-3.5 0l.328-.328c.578-.578.867-.867 1.235-1.02c.367-.152.776-.152 1.594-.152h.686c.818 0 1.226 0 1.594.152c.367.152.656.441 1.234 1.02L15.5 22m-13-10c0 2.828 0 4.243.879 5.121C4.257 18 5.672 18 8.5 18h7c2.828 0 4.243 0 5.121-.879c.879-.878.879-2.293.879-5.121V8c0-2.828 0-4.243-.879-5.121C19.743 2 18.328 2 15.5 2h-7c-2.828 0-4.243 0-5.121.879C2.5 3.757 2.5 5.172 2.5 8zM8 13v-3m4 3V7m4 6v-2"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function Check(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none" fillRule="evenodd">
-//         <path
-//           fill="currentColor"
-//           d="M21.546 5.111a1.5 1.5 0 0 1 0 2.121L10.303 18.475a1.6 1.6 0 0 1-2.263 0L2.454 12.89a1.5 1.5 0 1 1 2.121-2.121l4.596 4.596L19.424 5.111a1.5 1.5 0 0 1 2.122 0"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// export function CheckDouble(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M3 13.333S4.5 14 6.5 17c0 0 .285-.48.821-1.247M17 6c-2.291 1.146-4.688 3.552-6.612 5.822M8 13.333S9.5 14 11.5 17c0 0 5.5-8.5 10.5-11"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function ChevronDown(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M18 9s-4.419 6-6 6s-6-6-6-6"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function CloudUpload(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none" fillRule="evenodd">
-//         <path
-//           fill="currentColor"
-//           d="M11.5 4C8.851 4 6.739 6.38 7.027 9a1.01 1.01 0 0 1-.758 1.09A3.002 3.002 0 0 0 7 16h1a1 1 0 1 1 0 2H7a5 5 0 0 1-2-9.584a6.5 6.5 0 0 1 12.586-2.204A6.002 6.002 0 0 1 16 18a1 1 0 1 1 0-2a4 4 0 0 0 .655-7.947a1.01 1.01 0 0 1-.81-.732A4.5 4.5 0 0 0 11.5 4m1.5 8.416l1.293 1.292a1 1 0 0 0 1.414-1.416l-2.824-2.819a1.25 1.25 0 0 0-1.766 0l-2.824 2.82a1 1 0 0 0 1.414 1.415L11 12.416V21a1 1 0 1 0 2 0z"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function Close(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none" fillRule="evenodd">
-//         <path
-//           fill="currentColor"
-//           d="m12 13.414l5.657 5.657a1 1 0 0 0 1.414-1.414L13.414 12l5.657-5.657a1 1 0 0 0-1.414-1.414L12 10.586L6.343 4.929A1 1 0 0 0 4.93 6.343L10.586 12l-5.657 5.657a1 1 0 1 0 1.414 1.414z"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function CloseLight(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M19 5L5 19M5 5l14 14"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function Cog(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none" fillRule="evenodd">
-//         {/* <path d="m12.593 23.258l-.011.002l-.071.035l-.02.004l-.014-.004l-.071-.035q-.016-.005-.024.005l-.004.01l-.017.428l.005.02l.01.013l.104.074l.015.004l.012-.004l.104-.074l.012-.016l.004-.017l-.017-.427q-.004-.016-.017-.018m.265-.113l-.013.002l-.185.093l-.01.01l-.003.011l.018.43l.005.012l.008.007l.201.093q.019.005.029-.008l.004-.014l-.034-.614q-.005-.018-.02-.022m-.715.002a.02.02 0 0 0-.027.006l-.006.014l-.034.614q.001.018.017.024l.015-.002l.201-.093l.01-.008l.004-.011l.017-.43l-.003-.012l-.01-.01z"></path> */}
-//         <path
-//           fill="currentColor"
-//           d="M14.035 2.809c.37-.266.89-.39 1.401-.203a10 10 0 0 1 2.982 1.725c.417.35.57.861.524 1.313c-.075.753.057 1.48.42 2.106c.32.557.802.997 1.39 1.307l.225.11c.414.187.782.576.875 1.113a10 10 0 0 1 0 3.44c-.083.484-.39.847-.753 1.051l-.122.063c-.69.31-1.254.79-1.616 1.416c-.362.627-.494 1.353-.419 2.106c.045.452-.107.964-.524 1.313a10 10 0 0 1-2.982 1.725a1.51 1.51 0 0 1-1.4-.203C13.42 20.75 12.723 20.5 12 20.5s-1.42.249-2.035.691a1.51 1.51 0 0 1-1.401.203a10 10 0 0 1-2.982-1.725a1.51 1.51 0 0 1-.524-1.313c.075-.753-.058-1.48-.42-2.106a3.4 3.4 0 0 0-1.39-1.307l-.225-.11a1.51 1.51 0 0 1-.875-1.113a10 10 0 0 1 0-3.44c.083-.484.39-.847.753-1.051l.122-.062c.69-.311 1.254-.79 1.616-1.417c.361-.626.494-1.353.419-2.106a1.51 1.51 0 0 1 .524-1.313a10 10 0 0 1 2.982-1.725a1.51 1.51 0 0 1 1.4.203c.615.442 1.312.691 2.036.691s1.42-.249 2.035-.691m.957 1.769c-.866.57-1.887.922-2.992.922s-2.126-.353-2.992-.922A8 8 0 0 0 7.068 5.7c.06 1.033-.145 2.093-.697 3.05c-.553.956-1.368 1.663-2.293 2.128a8 8 0 0 0 0 2.242c.925.465 1.74 1.172 2.293 2.13c.552.955.757 2.015.697 3.048a8 8 0 0 0 1.94 1.123c.866-.57 1.887-.922 2.992-.922s2.126.353 2.992.922a8 8 0 0 0 1.94-1.122c-.06-1.034.145-2.094.697-3.05c.552-.957 1.368-1.664 2.293-2.13a8 8 0 0 0 0-2.24c-.925-.466-1.74-1.173-2.293-2.13c-.552-.956-.757-2.016-.697-3.05a8 8 0 0 0-1.94-1.122ZM12 8a4 4 0 1 1 0 8a4 4 0 0 1 0-8m0 2a2 2 0 1 0 0 4a2 2 0 0 0 0-4"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function Cocktail(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1"
-//         color="currentColor"
-//       >
-//         <path d="m8.205 15.358l-3.688-4.277C2.625 8.887 1.68 7.79 2.098 6.895S3.98 6 6.901 6h4.198c2.922 0 4.383 0 4.803.895c.42.894-.527 1.992-2.42 4.186l-3.687 4.277C9.425 15.786 9.24 16 9 16s-.426-.214-.795-.642M8.5 6l-.401-2.406a1 1 0 0 0-.67-.784L5 2m4 14v6m-1.5 0h3"></path>
-//         <path d="M15.86 8.833A3.52 3.52 0 0 0 18.482 10A3.51 3.51 0 0 0 22 6.5C22 4.567 20.425 3 18.482 3A3.51 3.51 0 0 0 15 6"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function CollectionIcon(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//       >
-//         <path
-//           fill="currentColor"
-//           d="M7.769 7.997c1.34-.822 2.51-.49 3.214.037c.288.216.432.325.517.325s.229-.109.517-.325c.703-.528 1.874-.86 3.214-.037c1.76 1.08 2.158 4.64-1.9 7.645c-.774.572-1.16.858-1.831.858c-.67 0-1.057-.286-1.83-.858c-4.06-3.005-3.661-6.566-1.901-7.645"
-//         ></path>
-//         <path d="M2 12c0-4.478 0-6.718 1.391-8.109S7.021 2.5 11.5 2.5c4.478 0 6.718 0 8.109 1.391S21 7.521 21 12c0 4.478 0 6.718-1.391 8.109S15.979 21.5 11.5 21.5c-4.478 0-6.718 0-8.109-1.391S2 16.479 2 12"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function Copy(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 16 16"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke={props.fill}
-//         strokeLinejoin="round"
-//         d="M12 10.5h1.5v-8h-8V4m-3 9.5h8v-8h-8z"
-//       ></path>
-//     </svg>
-//   );
-// }
-
-// function EGuitarFender(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 128 128"
-//       {...props}
-//     >
-//       <path
-//         fill="#ed6c30"
-//         d="M67.15 49.9s3.25 1.1 5.92-1.21l2.52 4.9l-7.04 5.79zm14.44 7.21s-8.18 8.91-5.48 13.46l-7.57-6.73l11.07-13.56z"
-//       ></path>
-//       <g fill="#2f2f2f">
-//         <path d="M108.41 18.19L70.36 51.04l4.91 13.15l38.21-42.8z"></path>
-//         <path d="M107.05 19.36s1.09-2.9 1.36-4.7c.26-1.8 2.93-3.8 5.07-5.13c1.48-.93 3.08-1.71 4.63-2.51c1.33-.68 2.77-1.31 4.28-1.39c3.25-.16 6.25 2.75 4.46 5.8c-.38.65-.71 1.28-1.2 1.87c-1.62 1.95-3.13 4-4.74 5.95c-.5.6-1.03.93-1.64 1.39c-1.27.96-2.75 1.62-4.36 1.7c-1.7.09-3.23 1.08-4.34 2.32z"></path>
-//         <path d="M109.91 12.49c-.12.1-.25.2-.37.3c-.06.05-.12.17-.2.19c.13-.04-.13-1.12-.22-1.24c-.46-.65-1.45-.18-2.07-.46c-1.25-.57-1.55-1.96-.99-3.15c.3-.66.87-1.15 1.57-1.33c.9-.23 1.88-.08 2.49.68c.72.9-.29 2.39.64 3.19c.11.1 1.07.3 1.15.18c-.04.05-.12.08-.17.12c-.09.07-.14.17-.23.24c-.23.19-.48.35-.72.54c-.29.24-.57.48-.88.74m4.45-4.18c-.01-.16-.04-.31-.09-.44c-.28-.72-1.43-.57-1.95-1.06c-.99-.93-.99-2.46-.06-3.43c.43-.46 1.02-.72 1.65-.74c.92-.02 1.85.34 2.27 1.21c.33.67.04 1.36-.17 2.01c-.18.56-.33 1.38.32 1.71c.07.03.44.16.45.26c0 .1-.32.24-.38.27c-.3.16-.59.34-.89.49c-.22.1-.45.18-.67.28c-.16.08-.39.24-.58.25c-.3.02-.01-.21.03-.33c.06-.12.08-.3.07-.48m5.67-2.33c.2-.86-.81-1.23-1.35-1.66c-1.11-.88-1.12-2.18-.18-3.2c.56-.61 1.38-.92 2.21-.77c.91.16 1.74.71 1.98 1.65c.17.65-.15 1.21-.47 1.75c-.15.26-.32.52-.41.81c-.09.31.04.6.19.88c.07.12.12.23.26.28c.08.03.17.01.23.06c-.32.14-.65.24-.99.34c-.3.09-.6.16-.91.22c-.15.03-.71.24-.83.15c.13-.14.22-.32.27-.51"></path>
-//       </g>
-//       <defs>
-//         <path
-//           id="notoV1Guitar0"
-//           d="M44.95 54.23C32.67 66.19 19.76 57.55 7.9 69.4c-14.22 14.21-6.03 37.47 10.85 50.3c10.85 8.24 28.12 11.81 39.72 2.36c6.87-5.59 9.91-14.49 11.67-22.87c.75-3.56.75-7.47 3.72-10.02c2.63-2.27 6.16-3.04 8.89-5.16c.23-.19.46-.38.68-.58c.26-.23.49-.48.7-.74c1.42-1.68 2.55-3.7 3.35-5.74c1.03-2.63 1.92-8.12-2.52-7.67c-.68.07-1.12.49-1.64.93c-1.2 1.02-2.45 2.13-4.15 1.96c-2.5-.25-3.87-2.06-3.94-4.43c-.03-1.19-.09-2.34.04-3.53c.12-1.14.81-2.07 1.08-3.2c.45-1.92 1.3-4.36.27-6.19c-.91-1.6-3.48-2.85-5.14-3.42c-1.83-.62-4.13-.96-5.57-2.36c-2.44-2.36-1.62-6.2-.31-8.91c.68-1.42 1.48-2.7 2.67-3.74c1.26-1.09 3.02-2.1 3.8-3.65c1.49-2.95-.19-5.36-3.27-5.34c-6.11.05-11.99 3.04-15.46 8.1c-2.12 3.09-2.28 7.33-3.5 10.81c-1.19 3.41-2.88 5.95-4.89 7.92"
-//         ></path>
-//       </defs>
-//       <use fill="#ed6c30" href="#notoV1Guitar0"></use>
-//       <clipPath id="notoV1Guitar1">
-//         <use href="#notoV1Guitar0"></use>
-//       </clipPath>
-//       <path
-//         fill="#fff"
-//         d="M60.51 57.27c-1.74 3.07-4.78 5.61-7.6 7.61c-1.78 1.26-3.77 2.17-5.74 3.07c-4.88 2.23-11.3 4.46-13.69 9.72c-3.33 7.3 3.83 12.37 8.53 16.7c1.29 1.19 2.58 2.51 3.59 3.94c1.78 2.49 2.96 5.39 4.14 8.18c1.43 3.4 4.16 7 8.14 7.28c9.89.7 8.37-16.88 10.87-22.7c.96-2.26 2.59-4.17 4.77-5.36c2.51-1.38 5.22-2.43 7.62-4c1.87-1.23 3.55-2.83 4.29-5c.39-1.12 1.58-4.57-.5-4.83c-1.05-.13-2 .96-2.85 1.46c-2.21 1.31-5.48 2.16-7.37-.15c-1.33-1.64-1.55-4.06-1.33-6.09c.28-2.56.52-4.26-.82-6.55c-.7-1.19-1.24-2.48-2.13-3.6c-1.33-1.68-3.05-3.65-5.21-4.22c-2.98-.8-3.43 2.13-4.48 4.13c-.08.15-.15.28-.23.41"
-//         clipPath="url(#notoV1Guitar1)"
-//       ></path>
-//       <path fill="#2f2f2f" d="m73.07 48.7l-8.1 6.84l9.25 9.6l8.53-9.25z"></path>
-//       <circle cx="57.29" cy="107.17" r="3.41" fill="#2f2f2f"></circle>
-//       <circle cx="54.48" cy="97.94" r="3.41" fill="#2f2f2f"></circle>
-//       <path
-//         fill="#2f2f2f"
-//         d="M56.64 64.58c-.79-.6-1.68-.87-2.78-.15c-3.13 2.07-.28 4.7 1.18 6.52q2.61 3.255 5.64 6.12c1.07 1.01 1.9 2.27 3.6 1.8c.94-.27 1.64-.84 2.01-1.75c.9-2.2-2.51-5.08-3.76-6.52c-1.12-1.27-2.3-2.5-3.44-3.74c-.76-.82-1.56-1.61-2.45-2.28m-9.71 8.1c-.78-.59-1.65-.86-2.7-.16c-3.02 1.99-.23 4.56 1.21 6.35c1.72 2.12 3.57 4.12 5.55 5.99c1.05.99 1.88 2.22 3.52 1.77c.93-.25 1.59-.8 1.94-1.69c.85-2.13-2.49-4.96-3.73-6.37c-1.09-1.25-2.25-2.44-3.38-3.66c-.75-.8-1.54-1.57-2.41-2.23m-5.51 14.74c-1.08-1.3-2.19-2.56-3.34-3.8c-.76-.81-1.56-1.59-2.45-2.27c-1.74-1.33-3.43.23-4.81 1.26c-2.94 2.2-.28 4.92 1.37 6.95c1.76 2.14 3.62 4.16 5.63 6.05c1.06 1 1.91 2.24 3.53 1.81c1.34-.35 2.59-1.4 3.39-2.5c1.17-1.65.27-2.94-.8-4.33c-.82-1.07-1.66-2.13-2.52-3.17"
-//       ></path>
-//     </svg>
-//   );
-// }
-
-// function EGuitarGray(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 32 32"
-//       {...props}
-//     >
-//       <g fill="currentColor">
-//         <path d="m13.735 19.496l-1.53-1.53a.61.61 0 0 0-.85 0a.61.61 0 0 0 0 .85l1.53 1.53c.23.23.61.23.85 0s.24-.61 0-.85m-3.54.49l1.53 1.53c.23.23.23.61 0 .85a.61.61 0 0 1-.85 0l-1.53-1.53a.61.61 0 0 1 0-.85a.61.61 0 0 1 .85 0"></path>
-//         <path d="M27.485 2.416a.683.683 0 0 0-.9-.36c-.35.15-.51.55-.36.9c.07.16.19.27.34.34c.13.06.24.15.3.28l.16.368l-.933.43l-.157-.358a.58.58 0 0 1-.01-.41c.04-.14.04-.31-.02-.47a.683.683 0 0 0-.9-.36c-.35.15-.51.55-.36.9c.07.16.19.27.34.34c.13.06.24.15.3.28l.164.375l-.943.435l-.16-.37a.53.53 0 0 1-.01-.41a.64.64 0 0 0-.02-.47a.683.683 0 0 0-.9-.36c-.35.15-.51.55-.36.9c.07.16.19.27.34.34c.13.06.24.15.3.28l.167.386l-.378.174a.78.78 0 0 0-.233 1.157l-.077-.077l-8.558 8.558a3.37 3.37 0 0 1 .728-3.588c.6-.61.63-1.59.03-2.2s-1.59-.61-2.19-.01c-.93.93-1.52 2.08-1.76 3.29c-.27 1.36-1.42 2.35-2.8 2.48c-1.66.15-3.27.86-4.53 2.13c-2.83 2.86-2.79 7.57.1 10.37c2.85 2.77 7.41 2.74 10.23-.08a7.3 7.3 0 0 0 2.1-4.53a3.1 3.1 0 0 1 2.49-2.77c.55-.11 1.09-.3 1.61-.56c.76-.38 1.07-1.31.69-2.07a1.546 1.546 0 0 0-2.07-.69a3.34 3.34 0 0 1-2.755.115l8.562-8.561a.78.78 0 0 0 1.143-.134l1.08-1.81c.12-.2.36-.29.58-.2c.57.22 1.24.1 1.68-.38c.54-.58.55-1.49.02-2.07c-.5-.549-1.206-.578-1.892-.293l-.158-.347a.58.58 0 0 1-.01-.41c.06-.15.06-.32-.01-.48m-14.28 12.78c.306.177.526.473.604.826l-.164.164c-.25.25-.25.65 0 .89l.99.99c.244.244.631.25.873.018c.29.078.735.141 1.023.182l.204.03q.037.007.07.01t.07.01l.017.002c.196.022.753.085 1.403-.042q.063-.011.128-.025c.367-.074.795-.16.982.375c.07.18.01.39-.25.56c-.15.1-.52.24-.69.28c-.82.19-1.4.52-1.94 1.11c-.72.68-1.16 1.5-1.32 2.39c-.12.68-.24 1.35-.52 1.97c-.3.64-.7 1.22-1.21 1.73c-.58.59-1.26 1.03-2.01 1.33c-.7.28-1.51-.05-1.78-.75c-.06-.15-.09-.32-.09-.48c.003-.252.063-.535.126-.832c.16-.751.338-1.594-.336-2.268l-2.25-2.25c-1-1-1.02-2.63-.03-3.65l.02-.02c.79-.8 1.83-1.21 2.92-1.25c.65-.03 1.21-.36 1.52-.87c.25-.41.71-.63 1.18-.58c.16.02.32.07.46.15m-3.84 9.57c.1.1.1.26-.01.35l-1.06 1.06c-.1.1-.26.1-.36 0l-2.41-2.41c-.1-.1-.1-.26 0-.36l1.07-1.05c.1-.1.26-.1.36 0z"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// export function Energy(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 32 32"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="m6.194 11.397l5.998-8.085c.47-.632 1.348-.239 1.348.603v6.258c0 .505.345.913.77.913h2.918c.663 0 1.016.927.578 1.518l-5.998 8.084c-.47.632-1.348.239-1.348-.603v-6.258c0-.505-.345-.913-.77-.913H6.771c-.663 0-1.016-.927-.578-1.517"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function CryptoCoinbase(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       aria-label="Coinbase logo"
-//       height="32"
-//       role="img"
-//       viewBox="0 0 48 48"
-//       width="32"
-//       xmlns="http://www.w3.org/2000/svg"
-//       {...props}
-//     >
-//       <title>Coinbase logo</title>
-//       <path
-//         d="M24,36c-6.63,0-12-5.37-12-12s5.37-12,12-12c5.94,0,10.87,4.33,11.82,10h12.09C46.89,9.68,36.58,0,24,0 C10.75,0,0,10.75,0,24s10.75,24,24,24c12.58,0,22.89-9.68,23.91-22H35.82C34.87,31.67,29.94,36,24,36z"
-//         fill="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function CryptoUSDC(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 32 32"
-//       {...props}
-//     >
-//       <g fill="none">
-//         <circle cx="16" cy="16" r="16" fill="#3E73C4"></circle>
-//         <g fill="#FFF">
-//           <path d="M20.022 18.124c0-2.124-1.28-2.852-3.84-3.156c-1.828-.243-2.193-.728-2.193-1.578s.61-1.396 1.828-1.396c1.097 0 1.707.364 2.011 1.275a.46.46 0 0 0 .427.303h.975a.416.416 0 0 0 .427-.425v-.06a3.04 3.04 0 0 0-2.743-2.489V9.142c0-.243-.183-.425-.487-.486h-.915c-.243 0-.426.182-.487.486v1.396c-1.829.242-2.986 1.456-2.986 2.974c0 2.002 1.218 2.791 3.778 3.095c1.707.303 2.255.668 2.255 1.639c0 .97-.853 1.638-2.011 1.638c-1.585 0-2.133-.667-2.316-1.578c-.06-.242-.244-.364-.427-.364h-1.036a.416.416 0 0 0-.426.425v.06c.243 1.518 1.219 2.61 3.23 2.914v1.457c0 .242.183.425.487.485h.915c.243 0 .426-.182.487-.485V21.34c1.829-.303 3.047-1.578 3.047-3.217z"></path>
-//           <path d="M12.892 24.497c-4.754-1.7-7.192-6.98-5.424-11.653c.914-2.55 2.925-4.491 5.424-5.402c.244-.121.365-.303.365-.607v-.85c0-.242-.121-.424-.365-.485c-.061 0-.183 0-.244.06a10.895 10.895 0 0 0-7.13 13.717c1.096 3.4 3.717 6.01 7.13 7.102c.244.121.488 0 .548-.243c.061-.06.061-.122.061-.243v-.85c0-.182-.182-.424-.365-.546m6.46-18.936c-.244-.122-.488 0-.548.242c-.061.061-.061.122-.061.243v.85c0 .243.182.485.365.607c4.754 1.7 7.192 6.98 5.424 11.653c-.914 2.55-2.925 4.491-5.424 5.402c-.244.121-.365.303-.365.607v.85c0 .242.121.424.365.485c.061 0 .183 0 .244-.06a10.895 10.895 0 0 0 7.13-13.717c-1.096-3.46-3.778-6.07-7.13-7.162"></path>
-//         </g>
-//       </g>
-//     </svg>
-//   );
-// }
-// function DotSm(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path fill="currentColor" d="M12 16a4 4 0 1 1 0-8a4 4 0 0 1 0 8Z"></path>
-//     </svg>
-//   );
-// }
-// function ImageIcon(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none">
-//         <path
-//           stroke="currentColor"
-//           strokeLinecap="round"
-//           strokeLinejoin="round"
-//           strokeWidth="1.5"
-//           d="M20.33 17.657c.11-.366.17-.755.17-1.157v-9a4 4 0 0 0-4-4h-9a4 4 0 0 0-4 4v9.07m16.83 1.087l-.088-.104l-2.466-2.976a2 2 0 0 0-3.073-.008l-1.312 1.566l-.214.261m7.153 1.26a4 4 0 0 1-3.713 2.842m0 0l-.117.002h-9a4 4 0 0 1-4-3.93m13.117 3.928l-.093-.106l-3.347-3.996m-9.676.175l.177-.201l3.206-3.827a2 2 0 0 1 3.066 0l3.227 3.853"
-//         ></path>
-//         <circle cx="15.091" cy="8.909" r="1.5" fill="currentColor"></circle>
-//       </g>
-//     </svg>
-//   );
-// }
-// function ImageGen(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         color="currentColor"
-//       >
-//         <path d="M20.998 11q.002.705.002 1.5c0 4.478 0 6.718-1.391 8.109S15.979 22 11.5 22c-4.478 0-6.718 0-8.109-1.391S2 16.979 2 12.5c0-4.478 0-6.718 1.391-8.109S7.021 3 11.5 3q.795 0 1.5.002"></path>
-//         <path d="m18.5 2l.258.697c.338.914.507 1.371.84 1.704c.334.334.791.503 1.705.841L22 5.5l-.697.258c-.914.338-1.371.507-1.704.84c-.334.334-.503.791-.841 1.705L18.5 9l-.258-.697c-.338-.914-.507-1.371-.84-1.704c-.334-.334-.791-.503-1.705-.841L15 5.5l.697-.258c.914-.338 1.371-.507 1.704-.84c.334-.334.503-.791.841-1.705zm-14 19.5c4.372-5.225 9.274-12.116 16.498-7.458"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function ImageUpload(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         color="currentColor"
-//       >
-//         <path d="M5 21c4.21-4.751 8.941-11.052 16-6.327M17 4.5c.491-.506 1.8-2.5 2.5-2.5M22 4.5c-.491-.506-1.8-2.5-2.5-2.5m0 0v8"></path>
-//         <path d="M21 13c-.002 4.147-.053 6.27-1.391 7.609C18.217 22 15.979 22 11.5 22c-4.478 0-6.718 0-8.109-1.391S2 16.979 2 12.5c0-4.478 0-6.718 1.391-8.109S7.021 3 11.5 3H14"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function InfoLine(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none">
-//         {/* <path d="m12.593 23.258l-.011.002l-.071.035l-.02.004l-.014-.004l-.071-.035q-.016-.005-.024.005l-.004.01l-.017.428l.005.02l.01.013l.104.074l.015.004l.012-.004l.104-.074l.012-.016l.004-.017l-.017-.427q-.004-.016-.017-.018m.265-.113l-.013.002l-.185.093l-.01.01l-.003.011l.018.43l.005.012l.008.007l.201.093q.019.005.029-.008l.004-.014l-.034-.614q-.005-.018-.02-.022m-.715.002a.02.02 0 0 0-.027.006l-.006.014l-.034.614q.001.018.017.024l.015-.002l.201-.093l.01-.008l.004-.011l.017-.43l-.003-.012l-.01-.01z"></path> */}
-//         <path
-//           stroke="none"
-//           fill={props.fill}
-//           d="M12 2c5.523 0 10 4.477 10 10s-4.477 10-10 10S2 17.523 2 12S6.477 2 12 2m0 2a8 8 0 1 0 0 16a8 8 0 0 0 0-16m-.01 6c.558 0 1.01.452 1.01 1.01v5.124A1 1 0 0 1 12.5 18h-.49A1.01 1.01 0 0 1 11 16.99V12a1 1 0 1 1 0-2zM12 7a1 1 0 1 1 0 2a1 1 0 0 1 0-2"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function EventStar(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         color="currentColor"
-//       >
-//         <path d="m17.777 13.65l.792 1.597a.98.98 0 0 0 .64.476l1.435.24c.918.155 1.134.826.472 1.489L20 18.577a.99.99 0 0 0-.234.82l.32 1.394c.252 1.102-.329 1.529-1.296.952l-1.345-.803c-.244-.145-.644-.145-.891 0l-1.346.803c-.963.577-1.548.146-1.296-.952l.32-1.393a.99.99 0 0 0-.234-.821l-1.116-1.125c-.657-.663-.445-1.334.472-1.488l1.436-.24a.98.98 0 0 0 .634-.477l.792-1.597c.432-.867 1.134-.867 1.561 0M18 2v2M6 2v2"></path>
-//         <path d="M21.5 11.5c-.004-3.866-.073-5.872-1.252-7.146C18.996 3 16.98 3 12.95 3h-1.9C7.02 3 5.004 3 3.752 4.354C2.5 5.707 2.5 7.886 2.5 12.244v.513c0 4.357 0 6.536 1.252 7.89c1.194 1.29 3.081 1.35 6.748 1.353M3 8h18"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function EyeOpen(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//       >
-//         <path d="M3.182 12.808C4.233 14.613 7.195 18.81 12 18.81c4.813 0 7.77-4.199 8.82-6.002a1.6 1.6 0 0 0-.001-1.615C19.769 9.389 16.809 5.19 12 5.19s-7.768 4.197-8.818 6.001a1.6 1.6 0 0 0 0 1.617Z"></path>
-//         <path d="M12 14.625a2.625 2.625 0 1 0 0-5.25a2.625 2.625 0 0 0 0 5.25Z"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function Fire(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none">
-//         {/* <path d="m12.593 23.258l-.011.002l-.071.035l-.02.004l-.014-.004l-.071-.035q-.016-.005-.024.005l-.004.01l-.017.428l.005.02l.01.013l.104.074l.015.004l.012-.004l.104-.074l.012-.016l.004-.017l-.017-.427q-.004-.016-.017-.018m.265-.113l-.013.002l-.185.093l-.01.01l-.003.011l.018.43l.005.012l.008.007l.201.093q.019.005.029-.008l.004-.014l-.034-.614q-.005-.018-.02-.022m-.715.002a.02.02 0 0 0-.027.006l-.006.014l-.034.614q.001.018.017.024l.015-.002l.201-.093l.01-.008l.004-.011l.017-.43l-.003-.012l-.01-.01z"></path> */}
-//         <path
-//           fill={props.fill}
-//           d="m11.514 2.142l-1.26-.755l-.24 1.449C9.632 5.124 8.069 7.25 6.345 8.744C2.97 11.67 2.231 14.85 3.276 17.475c1 2.512 3.538 4.232 6.114 4.519l.596.066c-1.474-.901-2.42-3.006-2.09-4.579c.326-1.546 1.438-2.994 3.574-4.33l1.077-.672l.402 1.205c.237.712.647 1.284 1.064 1.865c.2.28.403.563.589.864c.643 1.045.813 2.207.398 3.36c-.378 1.048-1.001 1.872-1.86 2.329l.97-.108c2.418-.269 4.193-1.096 5.346-2.479C20.599 18.144 21 16.379 21 14.5c0-1.75-.719-3.554-1.567-5.055c-.994-1.758-2.291-3.218-3.707-4.633c-.245.49-.226.688-.73 1.475a8.15 8.15 0 0 0-3.482-4.145"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function Fullscreen(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none" fillRule="evenodd">
-//         <path
-//           fill={props.fill}
-//           d="M18.5 5.5H16a1.5 1.5 0 0 1 0-3h3A2.5 2.5 0 0 1 21.5 5v3a1.5 1.5 0 0 1-3 0zM8 5.5H5.5V8a1.5 1.5 0 1 1-3 0V5A2.5 2.5 0 0 1 5 2.5h3a1.5 1.5 0 1 1 0 3m0 13H5.5V16a1.5 1.5 0 0 0-3 0v3A2.5 2.5 0 0 0 5 21.5h3a1.5 1.5 0 0 0 0-3m8 0h2.5V16a1.5 1.5 0 0 1 3 0v3a2.5 2.5 0 0 1-2.5 2.5h-3a1.5 1.5 0 0 1 0-3"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function Globe(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1"
-//         color="currentColor"
-//       >
-//         <circle cx="12" cy="12" r="10"></circle>
-//         <path d="M8 12c0 6 4 10 4 10s4-4 4-10s-4-10-4-10s-4 4-4 10m13 3H3m18-6H3"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function Heart(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill={"none"}
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M19.463 3.994c-2.682-1.645-5.023-.982-6.429.074c-.576.433-.864.65-1.034.65s-.458-.217-1.034-.65C9.56 3.012 7.219 2.349 4.537 3.994C1.018 6.153.222 13.274 8.34 19.284C9.886 20.427 10.659 21 12 21s2.114-.572 3.66-1.717c8.118-6.008 7.322-13.13 3.803-15.289"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function HeartFill(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="currentColor"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M19.463 3.994c-2.682-1.645-5.023-.982-6.429.074c-.576.433-.864.65-1.034.65s-.458-.217-1.034-.65C9.56 3.012 7.219 2.349 4.537 3.994C1.018 6.153.222 13.274 8.34 19.284C9.886 20.427 10.659 21 12 21s2.114-.572 3.66-1.717c8.118-6.008 7.322-13.13 3.803-15.289"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-
-// function HeartCheck(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="currentColor"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         color="currentColor"
-//       >
-//         <path d="M12 21c-1.341 0-2.114-.572-3.66-1.717C.222 13.275 1.018 6.153 4.537 3.994c2.682-1.645 5.023-.982 6.429.074c.576.433.864.65 1.034.65s.458-.217 1.034-.65c1.406-1.056 3.747-1.719 6.429-.074C21.18 5.048 22.25 7.286 21.949 10"></path>
-//         <path
-//           fill="currentColor"
-//           strokeWidth="3"
-//           color="#02C7BE"
-//           d="M14 18s1 0 2 2c0 0 3.177-5 6-6"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function HotelRoom(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         d="M1.5 18.5V14l.676-.184a37.34 37.34 0 0 1 19.648 0L22.5 14v4.5m-21 0s0 3-1.5 3m1.5-3h21m0 0s0 3 1.5 3M3.5 11c0-1.989-.297-3.966-.882-5.867L2.5 4.75V4.5h19v.25l-.118.383A20 20 0 0 0 20.5 11M12 7.5H6.5V11M12 7.5V11m0-3.5h5.5V11"
-//       ></path>
-//     </svg>
-//   );
-// }
-
-// function Info(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="currentColor"
-//         d="M12.3 7.29c.2-.18.44-.29.7-.29c.27 0 .5.11.71.29c.19.21.29.45.29.71c0 .27-.1.5-.29.71c-.21.19-.44.29-.71.29c-.26 0-.5-.1-.7-.29c-.19-.21-.3-.44-.3-.71c0-.26.11-.5.3-.71m-2.5 4.68s2.17-1.72 2.96-1.79c.74-.06.59.79.52 1.23l-.01.06c-.14.53-.31 1.17-.48 1.78c-.38 1.39-.75 2.75-.66 3c.1.34.72-.09 1.17-.39c.06-.04.11-.08.16-.11c0 0 .08-.08.16.03c.02.03.04.06.06.08c.09.14.14.19.02.27l-.04.02c-.22.15-1.16.81-1.54 1.05c-.41.27-1.98 1.17-1.74-.58c.21-1.23.49-2.29.71-3.12c.41-1.5.59-2.18-.33-1.59c-.37.22-.59.36-.72.45c-.11.08-.12.08-.19-.05l-.03-.06l-.05-.08c-.07-.1-.07-.11.03-.2M22 12c0 5.5-4.5 10-10 10S2 17.5 2 12S6.5 2 12 2s10 4.5 10 10m-2 0c0-4.42-3.58-8-8-8s-8 3.58-8 8s3.58 8 8 8s8-3.58 8-8"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function Launch(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="currentColor"
-//         d="m3.7 11.287l3.623-3.624q.293-.292.671-.413t.783-.04l1.415.294Q9.073 8.873 8.366 9.98t-1.414 2.688zm4.125 1.62q.652-1.51 1.563-2.89q.91-1.378 2.08-2.548q1.873-1.873 4.073-2.806t4.746-.797q.136 2.546-.795 4.746t-2.803 4.073q-1.164 1.163-2.548 2.07t-2.897 1.559zm6.17-2.768q.44.44 1.066.44t1.066-.44t.44-1.057t-.44-1.057t-1.066-.44t-1.067.44t-.44 1.057t.44 1.056m-1.161 10.314L11.444 17.2q1.581-.706 2.691-1.423q1.111-.717 2.48-1.836l.289 1.415q.08.404-.031.785q-.111.382-.404.675zm-7.687-4.306q.587-.586 1.423-.58t1.423.594q.587.586.587 1.423t-.587 1.423q-.51.51-1.635.873t-2.605.502q.138-1.479.511-2.602t.883-1.633"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function LeftChev(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none" fillRule="evenodd">
-//         <path
-//           fill="currentColor"
-//           d="M8.293 12.707a1 1 0 0 1 0-1.414l5.657-5.657a1 1 0 1 1 1.414 1.414L10.414 12l4.95 4.95a1 1 0 0 1-1.414 1.414z"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function MoreVertical(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none">
-//         {/* <path d="m12.593 23.258l-.011.002l-.071.035l-.02.004l-.014-.004l-.071-.035q-.016-.005-.024.005l-.004.01l-.017.428l.005.02l.01.013l.104.074l.015.004l.012-.004l.104-.074l.012-.016l.004-.017l-.017-.427q-.004-.016-.017-.018m.265-.113l-.013.002l-.185.093l-.01.01l-.003.011l.018.43l.005.012l.008.007l.201.093q.019.005.029-.008l.004-.014l-.034-.614q-.005-.018-.02-.022m-.715.002a.02.02 0 0 0-.027.006l-.006.014l-.034.614q.001.018.017.024l.015-.002l.201-.093l.01-.008l.004-.011l.017-.43l-.003-.012l-.01-.01z"></path> */}
-//         <path
-//           fill={props.fill}
-//           d="M12 16.5a1.5 1.5 0 1 1 0 3a1.5 1.5 0 0 1 0-3m0-6a1.5 1.5 0 1 1 0 3a1.5 1.5 0 0 1 0-3m0-6a1.5 1.5 0 1 1 0 3a1.5 1.5 0 0 1 0-3"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function Lock(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         color="currentColor"
-//       >
-//         <path d="M12 16.5v-2m-7.732 4.345c.225 1.67 1.608 2.979 3.292 3.056c1.416.065 2.855.099 4.44.099s3.024-.034 4.44-.1c1.684-.076 3.067-1.385 3.292-3.055c.147-1.09.268-2.207.268-3.345s-.121-2.255-.268-3.345c-.225-1.67-1.608-2.979-3.292-3.056A95 95 0 0 0 12 9c-1.585 0-3.024.034-4.44.1c-1.684.076-3.067 1.385-3.292 3.055C4.12 13.245 4 14.362 4 15.5s.121 2.255.268 3.345"></path>
-//         <path d="M7.5 9V6.5a4.5 4.5 0 0 1 9 0V9"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function MailSend(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         color="currentColor"
-//       >
-//         <path d="M22 12.5c0-.491-.005-1.483-.016-1.976c-.065-3.065-.098-4.598-1.229-5.733c-1.131-1.136-2.705-1.175-5.854-1.254a115 115 0 0 0-5.802 0c-3.149.079-4.723.118-5.854 1.254c-1.131 1.135-1.164 2.668-1.23 5.733a69 69 0 0 0 0 2.952c.066 3.065.099 4.598 1.23 5.733c1.131 1.136 2.705 1.175 5.854 1.254q1.204.03 2.401.036"></path>
-//         <path d="m7 8.5l2.942 1.74c1.715 1.014 2.4 1.014 4.116 0L17 8.5m5 9h-8m8 0c0-.7-1.994-2.008-2.5-2.5m2.5 2.5c0 .7-1.994 2.009-2.5 2.5"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function MapPin(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//       >
-//         <path d="M14.5 9a2.5 2.5 0 1 1-5 0a2.5 2.5 0 0 1 5 0m3.722 8c1.395 1.988 2.062 3.047 1.665 3.9q-.06.128-.14.247c-.575.853-2.06.853-5.03.853H9.283c-2.97 0-4.454 0-5.029-.853a2 2 0 0 1-.14-.247c-.397-.852.27-1.912 1.665-3.9"></path>
-//         <path d="M13.257 17.494a1.813 1.813 0 0 1-2.514 0c-3.089-2.993-7.228-6.336-5.21-11.19C6.626 3.679 9.246 2 12 2s5.375 1.68 6.467 4.304c2.016 4.847-2.113 8.207-5.21 11.19"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function MapPin2(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         color="currentColor"
-//       >
-//         <path d="M13.618 21.367A2.37 2.37 0 0 1 12 22a2.37 2.37 0 0 1-1.617-.633C6.412 17.626 1.09 13.447 3.685 7.38C5.09 4.1 8.458 2 12.001 2s6.912 2.1 8.315 5.38c2.592 6.06-2.717 10.259-6.698 13.987"></path>
-//         <path d="M15.5 11a3.5 3.5 0 1 1-7 0a3.5 3.5 0 0 1 7 0"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function Minus(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M20 12H4"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function MoreHorizontal(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none">
-//         <path
-//           fill={props.fill}
-//           d="M5 10a2 2 0 1 1 0 4a2 2 0 0 1 0-4m7 0a2 2 0 1 1 0 4a2 2 0 0 1 0-4m7 0a2 2 0 1 1 0 4a2 2 0 0 1 0-4"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function Mul(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         strokeWidth={0}
-//         fill="currentColor"
-//         d="M17.414 6.586a2 2 0 0 0-2.828 0L12 9.172L9.414 6.586a2 2 0 1 0-2.828 2.828L9.171 12l-2.585 2.586a2 2 0 1 0 2.828 2.828L12 14.828l2.586 2.586c.39.391.902.586 1.414.586s1.024-.195 1.414-.586a2 2 0 0 0 0-2.828L14.829 12l2.585-2.586a2 2 0 0 0 0-2.828"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function NumberList(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M11 6h10m-10 6h10m-10 6h10M3 15h1.5c.279 0 .418 0 .534.023a1.2 1.2 0 0 1 .943.943C6 16.082 6 16.22 6 16.5s0 .418-.023.534a1.2 1.2 0 0 1-.943.943C4.918 18 4.78 18 4.5 18s-.418 0-.534.023a1.2 1.2 0 0 0-.943.943C3 19.082 3 19.22 3 19.5v.9c0 .283 0 .424.088.512s.23.088.512.088H6M3 3h1.2a.3.3 0 0 1 .3.3V9m0 0H3m1.5 0H6"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function PayMongo(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 32 32"
-//       {...props}
-//     >
-//       <path
-//         d="M30 15a15 15 0 01-13.83 15A24.25 24.25 0 0115 22.53v-3.74a9.36 9.36 0 003.75 7.49 15 15 0 004.54-5C26.92 14.86 20.61 5.92 15 1.87c-5.61 4-11.93 13-8.3 19.43a15 15 0 007 6.37L13.32 30A15 15 0 1130 15z"
-//         fillRule="evenodd"
-//         fill="#22B47E"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function Pen(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="m15.214 5.982l1.402-1.401a1.982 1.982 0 0 1 2.803 2.803l-1.401 1.402m-2.804-2.804L6.98 14.216c-1.045 1.046-1.568 1.568-1.924 2.205S4.342 18.561 4 20c1.438-.342 2.942-.7 3.579-1.056s1.16-.879 2.205-1.924l8.234-8.234m-2.804-2.804l2.804 2.804M11 20h6"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function PencilEdit(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         color="currentColor"
-//       >
-//         <path d="m16.214 4.982l1.402-1.401a1.982 1.982 0 0 1 2.803 2.803l-1.401 1.402m-2.804-2.804l-5.234 5.234c-1.045 1.046-1.568 1.568-1.924 2.205S8.342 14.561 8 16c1.438-.342 2.942-.7 3.579-1.056s1.16-.879 2.205-1.924l5.234-5.234m-2.804-2.804l2.804 2.804"></path>
-//         <path d="M21 12c0 4.243 0 6.364-1.318 7.682S16.242 21 12 21s-6.364 0-7.682-1.318S3 16.242 3 12s0-6.364 1.318-7.682S7.758 3 12 3"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function Play(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M18.89 12.846c-.353 1.343-2.023 2.292-5.364 4.19c-3.23 1.835-4.845 2.752-6.146 2.384a3.25 3.25 0 0 1-1.424-.841C5 17.614 5 15.743 5 12s0-5.614.956-6.579a3.25 3.25 0 0 1 1.424-.84c1.301-.37 2.916.548 6.146 2.383c3.34 1.898 5.011 2.847 5.365 4.19a3.3 3.3 0 0 1 0 1.692"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function Plus(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M12 4v16m8-8H4"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// export function PlusMinus(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M2.5 12c0-4.478 0-6.718 1.391-8.109S7.521 2.5 12 2.5c4.478 0 6.718 0 8.109 1.391S21.5 7.521 21.5 12c0 4.478 0 6.718-1.391 8.109S16.479 21.5 12 21.5c-4.478 0-6.718 0-8.109-1.391S2.5 16.479 2.5 12M7 17L17 7M9 7v2m0 0v2m0-2h2M9 9H7m10 6h-4"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function Printer(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 16 16"
-//       {...props}
-//     >
-//       <path
-//         fill={props.fill}
-//         d="M14.25 6.57h-.75v-.05a1.26 1.26 0 0 0-.37-.89L9.37 1.87a1.26 1.26 0 0 0-.89-.37H3.75A1.25 1.25 0 0 0 2.5 2.75v3.82h-.75A1.26 1.26 0 0 0 .5 7.85v3.84A1.26 1.26 0 0 0 1.75 13h.75v.28a1.25 1.25 0 0 0 1.25 1.22h8.5a1.25 1.25 0 0 0 1.25-1.25V13h.75a1.26 1.26 0 0 0 1.25-1.28V7.85a1.26 1.26 0 0 0-1.25-1.28M3.75 2.75h4.73l3.77 3.77v.69h-8.5zm8.5 10.5h-8.5v-.92h8.5zm2-1.56h-.75v-.64h-11v.64h-.74V7.85h.74v.64h11v-.67h.74z"
-//       ></path>
-//       <ellipse
-//         cx="12.88"
-//         cy="9.77"
-//         fill={props.fill}
-//         rx=".63"
-//         ry=".59"
-//       ></ellipse>
-//     </svg>
-//   );
-// }
-
-// function Printer2(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         color="currentColor"
-//       >
-//         <path d="M7.354 18c-2.123 0-3.185 0-3.94-.453a3.04 3.04 0 0 1-1.15-1.223c-.392-.77-.287-1.787-.075-3.822c.176-1.698.264-2.547.698-3.171c.285-.41.67-.745 1.121-.977C4.695 8 5.582 8 7.354 8h9.292c1.772 0 2.659 0 3.346.354c.451.232.836.567 1.121.977c.434.624.522 1.473.698 3.172c.212 2.034.317 3.052-.076 3.821a3.04 3.04 0 0 1-1.148 1.223C19.83 18 18.769 18 16.646 18M17 8V6c0-1.886 0-2.828-.586-3.414S14.886 2 13 2h-2c-1.886 0-2.828 0-3.414.586S7 4.114 7 6v2"></path>
-//         <path d="M13.989 16H10.01c-.685 0-1.028 0-1.32.109a1.87 1.87 0 0 0-.945.8c-.168.281-.251.642-.417 1.363c-.26 1.128-.39 1.691-.301 2.143c.117.602.484 1.112.995 1.382c.382.203.918.203 1.988.203h3.978c1.07 0 1.606 0 1.988-.203c.51-.27.878-.78.995-1.382c.089-.452-.041-1.015-.3-2.143c-.167-.72-.25-1.082-.418-1.362a1.87 1.87 0 0 0-.946-.801C15.017 16 14.674 16 13.988 16M18 12h.009"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function QrCode(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M3 6c0-1.414 0-2.121.44-2.56C3.878 3 4.585 3 6 3s2.121 0 2.56.44C9 3.878 9 4.585 9 6s0 2.121-.44 2.56C8.122 9 7.415 9 6 9s-2.121 0-2.56-.44C3 8.122 3 7.415 3 6m0 12c0-1.414 0-2.121.44-2.56C3.878 15 4.585 15 6 15s2.121 0 2.56.44C9 15.878 9 16.585 9 18s0 2.121-.44 2.56C8.122 21 7.415 21 6 21s-2.121 0-2.56-.44C3 20.122 3 19.415 3 18m0-6h6m3-9v5m3-2c0-1.414 0-2.121.44-2.56C15.878 3 16.585 3 18 3s2.121 0 2.56.44C21 3.878 21 4.585 21 6s0 2.121-.44 2.56C20.122 9 19.415 9 18 9s-2.121 0-2.56-.44C15 8.122 15 7.415 15 6m6 6h-6c-1.414 0-2.121 0-2.56.44C12 12.878 12 13.585 12 15m0 2.77v2.768M15 15v1.5c0 1.446.784 1.5 2 1.5a1 1 0 0 1 1 1m-2 2h-1m3-6c1.414 0 2.121 0 2.56.44s.44 1.148.44 2.564s0 2.125-.44 2.565c-.32.32-.783.408-1.56.431"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function RightChev(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none" fillRule="evenodd">
-//         <path
-//           fill={props.fill}
-//           d="M16.06 10.94a1.5 1.5 0 0 1 0 2.12l-5.656 5.658a1.5 1.5 0 1 1-2.121-2.122L12.879 12L8.283 7.404a1.5 1.5 0 0 1 2.12-2.122l5.658 5.657Z"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function RightLine(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none" fillRule="evenodd">
-//         <path
-//           fill={props.fill}
-//           d="M15.707 11.293a1 1 0 0 1 0 1.414l-5.657 5.657a1 1 0 1 1-1.414-1.414l4.95-4.95l-4.95-4.95a1 1 0 0 1 1.414-1.414z"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function RightLineSm(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none" fillRule="evenodd">
-//         <path
-//           fill={props.fill}
-//           d="M10.293 8.465a1 1 0 0 0 0 1.414L12.414 12l-2.121 2.121a1 1 0 0 0 1.414 1.415l2.829-2.829a1 1 0 0 0 0-1.414l-2.829-2.828a1 1 0 0 0-1.414 0"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function Screen(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none">
-//         <path
-//           fill="currentColor"
-//           d="M11.514 16.903a.6.6 0 0 1 .976 0l2.766 3.868a.6.6 0 0 1-.488.949H9.237a.6.6 0 0 1-.488-.949ZM20 4a2 2 0 0 1 2 2v11a2 2 0 0 1-2 2h-3.71l-3.151-4.407a1.4 1.4 0 0 0-2.278 0L7.711 19H4a2 2 0 0 1-2-2V6a2 2 0 0 1 2-2z"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function Search(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none" fillRule="evenodd">
-//         <path
-//           fill={props.fill}
-//           d="M10.5 2a8.5 8.5 0 1 0 5.262 15.176l3.652 3.652a1 1 0 0 0 1.414-1.414l-3.652-3.652A8.5 8.5 0 0 0 10.5 2M4 10.5a6.5 6.5 0 1 1 13 0a6.5 6.5 0 0 1-13 0"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function Settings(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M7.05 6.462a2 2 0 0 0 2.63-1.519l.32-1.72a9 9 0 0 1 3.998 0l.322 1.72a2 2 0 0 0 2.63 1.519l1.649-.58a9 9 0 0 1 2.001 3.46l-1.33 1.14a2 2 0 0 0 0 3.037l1.33 1.139a9 9 0 0 1-2.001 3.46l-1.65-.58a2 2 0 0 0-2.63 1.519L14 20.777a9 9 0 0 1-3.998 0l-.322-1.72a2 2 0 0 0-2.63-1.519l-1.649.58a9 9 0 0 1-2.001-3.46l1.33-1.14a2 2 0 0 0 0-3.036L3.4 9.342a9 9 0 0 1 2-3.46zM12 9a3 3 0 1 1 0 6a3 3 0 0 1 0-6"
-//         clipRule="evenodd"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function Settings2(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1"
-//         color="currentColor"
-//       >
-//         <path d="M2.5 12c0-4.478 0-6.718 1.391-8.109S7.521 2.5 12 2.5c4.478 0 6.718 0 8.109 1.391S21.5 7.521 21.5 12c0 4.478 0 6.718-1.391 8.109S16.479 21.5 12 21.5c-4.478 0-6.718 0-8.109-1.391S2.5 16.479 2.5 12"></path>
-//         <path d="M8.5 10a1.5 1.5 0 1 1 0-3a1.5 1.5 0 0 1 0 3m7 7a1.5 1.5 0 1 0 0-3a1.5 1.5 0 0 0 0 3M10 8.5h7m-3 7H7"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function Share(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M19.192 9.441L17.227 7.47c-1.417-1.421-1.972-2.185-2.738-1.915c-.957.337-.642 2.463-.642 3.18c-1.486 0-3.032-.131-4.497.144C4.513 9.787 3 13.715 3 18c1.37-.97 2.737-2.003 4.382-2.452c2.054-.562 4.348-.294 6.465-.294c0 .718-.314 2.844.642 3.181c.868.306 1.321-.494 2.738-1.915l1.965-1.972C20.397 13.338 21 12.733 21 11.995c0-.74-.603-1.344-1.808-2.554"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-
-// function ShieldCheck(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 16 16"
-//       {...props}
-//     >
-//       <g fill="currentColor">
-//         <path d="M5.338 1.59a61 61 0 0 0-2.837.856a.48.48 0 0 0-.328.39c-.554 4.157.726 7.19 2.253 9.188a10.7 10.7 0 0 0 2.287 2.233c.346.244.652.42.893.533q.18.085.293.118a1 1 0 0 0 .101.025a1 1 0 0 0 .1-.025q.114-.034.294-.118c.24-.113.547-.29.893-.533a10.7 10.7 0 0 0 2.287-2.233c1.527-1.997 2.807-5.031 2.253-9.188a.48.48 0 0 0-.328-.39c-.651-.213-1.75-.56-2.837-.855C9.552 1.29 8.531 1.067 8 1.067c-.53 0-1.552.223-2.662.524zM5.072.56C6.157.265 7.31 0 8 0s1.843.265 2.928.56c1.11.3 2.229.655 2.887.87a1.54 1.54 0 0 1 1.044 1.262c.596 4.477-.787 7.795-2.465 9.99a11.8 11.8 0 0 1-2.517 2.453a7 7 0 0 1-1.048.625c-.28.132-.581.24-.829.24s-.548-.108-.829-.24a7 7 0 0 1-1.048-.625a11.8 11.8 0 0 1-2.517-2.453C1.928 10.487.545 7.169 1.141 2.692A1.54 1.54 0 0 1 2.185 1.43A63 63 0 0 1 5.072.56"></path>
-//         <path d="M10.854 5.146a.5.5 0 0 1 0 .708l-3 3a.5.5 0 0 1-.708 0l-1.5-1.5a.5.5 0 1 1 .708-.708L7.5 7.793l2.646-2.647a.5.5 0 0 1 .708 0"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function Sparkle(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M2 12q9 1 10 10q1-9 10-10q-9-1-10-10q-1 9-10 10"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function SidebarLeft(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none">
-//         <path
-//           fill={props.fill}
-//           d="M8 3v18H5a2 2 0 0 1-1.995-1.85L3 19V5a2 2 0 0 1 1.85-1.995L5 3zm11 0a2 2 0 0 1 1.995 1.85L21 5v14a2 2 0 0 1-1.85 1.995L19 21h-9V3z"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// export function Sparkles(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1"
-//         d="m15 2l.539 2.392a5.39 5.39 0 0 0 4.07 4.07L22 9l-2.392.539a5.39 5.39 0 0 0-4.07 4.07L15 16l-.539-2.392a5.39 5.39 0 0 0-4.07-4.07L8 9l2.392-.539a5.39 5.39 0 0 0 4.07-4.07zM7 12l.385 1.708a3.85 3.85 0 0 0 2.907 2.907L12 17l-1.708.385a3.85 3.85 0 0 0-2.907 2.907L7 22l-.385-1.708a3.85 3.85 0 0 0-2.907-2.907L2 17l1.708-.385a3.85 3.85 0 0 0 2.907-2.907z"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-
-// function Sparkles2(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none">
-//         <path
-//           strokeWidth={0}
-//           fill="currentColor"
-//           d="m6.173 13.004l.17.371a8.7 8.7 0 0 0 3.3 3.638l.265.155a.181.181 0 0 1 0 .316a8.7 8.7 0 0 0-3.565 3.793l-.17.372a.19.19 0 0 1-.346 0l-.17-.372a8.7 8.7 0 0 0-3.565-3.793a.181.181 0 0 1 0-.316l.265-.155a8.7 8.7 0 0 0 3.3-3.638l.17-.371a.19.19 0 0 1 .346 0M15.078 2.73a15.2 15.2 0 0 0 6.536 7.288a.317.317 0 0 1 0 .553a15.2 15.2 0 0 0-6.536 7.289a.331.331 0 0 1-.605 0a15.2 15.2 0 0 0-6.536-7.289a.317.317 0 0 1 0-.553a15.2 15.2 0 0 0 6.536-7.288a.331.331 0 0 1 .605 0"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function SpinnerBall(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <ellipse cx="12" cy="5" fill="currentColor" rx="4" ry="4">
-//         <animate
-//           id="svgSpinnersBouncingBall0"
-//           fill="freeze"
-//           attributeName="cy"
-//           begin="0;svgSpinnersBouncingBall2.end"
-//           calcMode="spline"
-//           dur="0.375s"
-//           keySplines=".33,0,.66,.33"
-//           values="5;20"
-//         ></animate>
-//         <animate
-//           attributeName="rx"
-//           begin="svgSpinnersBouncingBall0.end"
-//           calcMode="spline"
-//           dur="0.05s"
-//           keySplines=".33,0,.66,.33;.33,.66,.66,1"
-//           values="4;4.8;4"
-//         ></animate>
-//         <animate
-//           attributeName="ry"
-//           begin="svgSpinnersBouncingBall0.end"
-//           calcMode="spline"
-//           dur="0.05s"
-//           keySplines=".33,0,.66,.33;.33,.66,.66,1"
-//           values="4;3;4"
-//         ></animate>
-//         <animate
-//           id="svgSpinnersBouncingBall1"
-//           attributeName="cy"
-//           begin="svgSpinnersBouncingBall0.end"
-//           calcMode="spline"
-//           dur="0.025s"
-//           keySplines=".33,0,.66,.33"
-//           values="20;20.5"
-//         ></animate>
-//         <animate
-//           id="svgSpinnersBouncingBall2"
-//           attributeName="cy"
-//           begin="svgSpinnersBouncingBall1.end"
-//           calcMode="spline"
-//           dur="0.4s"
-//           keySplines=".33,.66,.66,1"
-//           values="20.5;5"
-//         ></animate>
-//       </ellipse>
-//     </svg>
-//   );
-// }
-// function SpinnerClock(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         strokeWidth="0"
-//         fill="#ccc"
-//         d="M12,1A11,11,0,1,0,23,12,11,11,0,0,0,12,1Zm0,20a9,9,0,1,1,9-9A9,9,0,0,1,12,21Z"
-//       ></path>
-//       <rect width="1.2" height="6.5" x="11.5" y="6" fill="red" rx="1">
-//         <animateTransform
-//           attributeName="transform"
-//           dur="9s"
-//           repeatCount="indefinite"
-//           type="rotate"
-//           values="0 12 12;360 12 12"
-//         ></animateTransform>
-//       </rect>
-//       <rect
-//         width="0.6"
-//         height="7.5"
-//         x="11.5"
-//         y="11.5"
-//         fill="currentColor"
-//         rx="3"
-//       >
-//         <animateTransform
-//           attributeName="transform"
-//           dur="0.75s"
-//           repeatCount="indefinite"
-//           type="rotate"
-//           values="0 12 12;360 12 12"
-//         ></animateTransform>
-//       </rect>
-//     </svg>
-//   );
-// }
-
-// function SpinnersPulseRing(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="currentColor"
-//         d="M12,1A11,11,0,1,0,23,12,11,11,0,0,0,12,1Zm0,20a9,9,0,1,1,9-9A9,9,0,0,1,12,21Z"
-//         transform="matrix(0 0 0 0 12 12)"
-//       >
-//         <animateTransform
-//           attributeName="transform"
-//           calcMode="spline"
-//           dur="1.2s"
-//           keySplines=".52,.6,.25,.99"
-//           repeatCount="indefinite"
-//           type="translate"
-//           values="12 12;0 0"
-//         ></animateTransform>
-//         <animateTransform
-//           additive="sum"
-//           attributeName="transform"
-//           calcMode="spline"
-//           dur="1.2s"
-//           keySplines=".52,.6,.25,.99"
-//           repeatCount="indefinite"
-//           type="scale"
-//           values="0;1"
-//         ></animateTransform>
-//         <animate
-//           attributeName="opacity"
-//           calcMode="spline"
-//           dur="1.2s"
-//           keySplines=".52,.6,.25,.99"
-//           repeatCount="indefinite"
-//           values="1;0"
-//         ></animate>
-//       </path>
-//     </svg>
-//   );
-// }
-// function SpinnerDotScale(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <circle cx="4" cy="12" r="3" fill="currentColor">
-//         <animate
-//           id="svgSpinners3DotsScale0"
-//           attributeName="r"
-//           begin="0;svgSpinners3DotsScale1.end-0.25s"
-//           dur="0.75s"
-//           values="3;.2;3"
-//         ></animate>
-//       </circle>
-//       <circle cx="12" cy="12" r="3" fill="currentColor">
-//         <animate
-//           attributeName="r"
-//           begin="svgSpinners3DotsScale0.end-0.6s"
-//           dur="0.75s"
-//           values="3;.2;3"
-//         ></animate>
-//       </circle>
-//       <circle cx="20" cy="12" r="3" fill="currentColor">
-//         <animate
-//           id="svgSpinners3DotsScale1"
-//           attributeName="r"
-//           begin="svgSpinners3DotsScale0.end-0.45s"
-//           dur="0.75s"
-//           values="3;.2;3"
-//         ></animate>
-//       </circle>
-//     </svg>
-//   );
-// }
-// function SpinnerDotMove(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <circle cx="4" cy="12" r="0" fill="currentColor">
-//         <animate
-//           fill="freeze"
-//           attributeName="r"
-//           begin="0;svgSpinners3DotsMove1.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="0;3"
-//         ></animate>
-//         <animate
-//           fill="freeze"
-//           attributeName="cx"
-//           begin="svgSpinners3DotsMove7.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="4;12"
-//         ></animate>
-//         <animate
-//           fill="freeze"
-//           attributeName="cx"
-//           begin="svgSpinners3DotsMove5.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="12;20"
-//         ></animate>
-//         <animate
-//           id="svgSpinners3DotsMove0"
-//           fill="freeze"
-//           attributeName="r"
-//           begin="svgSpinners3DotsMove3.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="3;0"
-//         ></animate>
-//         <animate
-//           id="svgSpinners3DotsMove1"
-//           fill="freeze"
-//           attributeName="cx"
-//           begin="svgSpinners3DotsMove0.end"
-//           dur="0.001s"
-//           values="20;4"
-//         ></animate>
-//       </circle>
-//       <circle cx="4" cy="12" r="3" fill="currentColor">
-//         <animate
-//           fill="freeze"
-//           attributeName="cx"
-//           begin="0;svgSpinners3DotsMove1.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="4;12"
-//         ></animate>
-//         <animate
-//           fill="freeze"
-//           attributeName="cx"
-//           begin="svgSpinners3DotsMove7.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="12;20"
-//         ></animate>
-//         <animate
-//           id="svgSpinners3DotsMove2"
-//           fill="freeze"
-//           attributeName="r"
-//           begin="svgSpinners3DotsMove5.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="3;0"
-//         ></animate>
-//         <animate
-//           id="svgSpinners3DotsMove3"
-//           fill="freeze"
-//           attributeName="cx"
-//           begin="svgSpinners3DotsMove2.end"
-//           dur="0.001s"
-//           values="20;4"
-//         ></animate>
-//         <animate
-//           fill="freeze"
-//           attributeName="r"
-//           begin="svgSpinners3DotsMove3.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="0;3"
-//         ></animate>
-//       </circle>
-//       <circle cx="12" cy="12" r="3" fill="currentColor">
-//         <animate
-//           fill="freeze"
-//           attributeName="cx"
-//           begin="0;svgSpinners3DotsMove1.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="12;20"
-//         ></animate>
-//         <animate
-//           id="svgSpinners3DotsMove4"
-//           fill="freeze"
-//           attributeName="r"
-//           begin="svgSpinners3DotsMove7.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="3;0"
-//         ></animate>
-//         <animate
-//           id="svgSpinners3DotsMove5"
-//           fill="freeze"
-//           attributeName="cx"
-//           begin="svgSpinners3DotsMove4.end"
-//           dur="0.001s"
-//           values="20;4"
-//         ></animate>
-//         <animate
-//           fill="freeze"
-//           attributeName="r"
-//           begin="svgSpinners3DotsMove5.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="0;3"
-//         ></animate>
-//         <animate
-//           fill="freeze"
-//           attributeName="cx"
-//           begin="svgSpinners3DotsMove3.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="4;12"
-//         ></animate>
-//       </circle>
-//       <circle cx="20" cy="12" r="3" fill="currentColor">
-//         <animate
-//           id="svgSpinners3DotsMove6"
-//           fill="freeze"
-//           attributeName="r"
-//           begin="0;svgSpinners3DotsMove1.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="3;0"
-//         ></animate>
-//         <animate
-//           id="svgSpinners3DotsMove7"
-//           fill="freeze"
-//           attributeName="cx"
-//           begin="svgSpinners3DotsMove6.end"
-//           dur="0.001s"
-//           values="20;4"
-//         ></animate>
-//         <animate
-//           fill="freeze"
-//           attributeName="r"
-//           begin="svgSpinners3DotsMove7.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="0;3"
-//         ></animate>
-//         <animate
-//           fill="freeze"
-//           attributeName="cx"
-//           begin="svgSpinners3DotsMove5.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="4;12"
-//         ></animate>
-//         <animate
-//           fill="freeze"
-//           attributeName="cx"
-//           begin="svgSpinners3DotsMove3.end"
-//           calcMode="spline"
-//           dur="0.5s"
-//           keySplines=".36,.6,.31,1"
-//           values="12;20"
-//         ></animate>
-//       </circle>
-//     </svg>
-//   );
-// }
-// function Stats(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none" fillRule="evenodd">
-//         <path
-//           fill="currentColor"
-//           d="M2 4a1 1 0 0 1 1-1h18a1 1 0 1 1 0 2v11a2 2 0 0 1-2 2h-5.055l2.293 2.293a1 1 0 0 1-1.414 1.414l-2.828-2.828l-2.829 2.828a1 1 0 0 1-1.414-1.414L10.046 18H5a2 2 0 0 1-2-2V5a1 1 0 0 1-1-1m14.243 3.172a1 1 0 0 1 1.414 1.414l-4.236 4.236a1.01 1.01 0 0 1-1.428 0L9.88 10.709l-2.121 2.12a1 1 0 0 1-1.415-1.413l2.822-2.822a1.01 1.01 0 0 1 1.428 0l2.113 2.113z"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function Squircle(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="0"
-//         stroke="currentColor"
-//         d="M12 3c7.2 0 9 1.8 9 9s-1.8 9-9 9s-9-1.8-9-9s1.8-9 9-9"
-//       ></path>
-//     </svg>
-//   );
-// }
-
-// function Support(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         color="currentColor"
-//       >
-//         <path d="M17 10.805c0-.346 0-.519.052-.673c.151-.448.55-.621.95-.803c.448-.205.672-.307.895-.325c.252-.02.505.034.721.155c.286.16.486.466.69.714c.943 1.146 1.415 1.719 1.587 2.35c.14.51.14 1.044 0 1.553c-.251.922-1.046 1.694-1.635 2.41c-.301.365-.452.548-.642.655a1.27 1.27 0 0 1-.721.155c-.223-.018-.447-.12-.896-.325c-.4-.182-.798-.355-.949-.803c-.052-.154-.052-.327-.052-.673zm-10 0c0-.436-.012-.827-.364-1.133c-.128-.111-.298-.188-.637-.343c-.449-.204-.673-.307-.896-.325c-.667-.054-1.026.402-1.41.87c-.944 1.145-1.416 1.718-1.589 2.35a2.94 2.94 0 0 0 0 1.553c.252.921 1.048 1.694 1.636 2.409c.371.45.726.861 1.363.81c.223-.018.447-.12.896-.325c.34-.154.509-.232.637-.343c.352-.306.364-.697.364-1.132z"></path>
-//         <path d="M5 9c0-3.314 3.134-6 7-6s7 2.686 7 6m0 8v.8c0 1.767-1.79 3.2-4 3.2h-2"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function Tablet(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//       >
-//         <rect width="18.5" height="15" x="2.75" y="4.5" rx="3"></rect>
-//         <path d="M10 16h4"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// export function Ticket(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M19.709 7.65c.157-.157.428-.173.587 0c.932 1.018 1.449 1.772 1.619 2.606c.097.48.111.968.04 1.443c-.19 1.285-1.23 2.325-3.312 4.406l-2.538 2.538c-2.08 2.081-3.121 3.122-4.406 3.313c-.475.07-.963.056-1.443-.041c-.834-.17-1.588-.686-2.605-1.619c-.174-.16-.158-.43 0-.588c.876-.876.834-2.338-.093-3.266c-.928-.927-2.39-.969-3.266-.093c-.157.158-.429.174-.588 0c-.933-1.017-1.449-1.77-1.619-2.605a4.2 4.2 0 0 1-.04-1.443c.19-1.285 1.23-2.325 3.312-4.406l2.538-2.538c2.08-2.081 3.121-3.122 4.406-3.313c.475-.07.963-.056 1.443.041c.835.17 1.588.687 2.605 1.62c.174.159.158.43 0 .586c-.876.877-.834 2.339.094 3.266c.927.928 2.39.97 3.265.093M19 15L9 5"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function Ticket2(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M2.464 9.344c-.248 0-.475-.202-.464-.465c.067-1.542.255-2.546.78-3.34a4.7 4.7 0 0 1 1.11-1.174C5.055 3.5 6.7 3.5 9.992 3.5h4.014c3.292 0 4.937 0 6.104.865c.431.32.807.717 1.109 1.174c.525.794.713 1.798.78 3.34c.011.263-.216.465-.465.465c-1.386 0-2.51 1.189-2.51 2.656s1.124 2.656 2.51 2.656c.249 0 .476.202.465.466c-.067 1.541-.255 2.545-.78 3.34a4.7 4.7 0 0 1-1.11 1.173c-1.166.865-2.811.865-6.103.865H9.993c-3.292 0-4.937 0-6.104-.865a4.7 4.7 0 0 1-1.109-1.174c-.525-.794-.713-1.798-.78-3.34c-.011-.263.216-.465.464-.465c1.386 0 2.51-1.189 2.51-2.656S3.85 9.344 2.464 9.344M9 3.5v17"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function TicketFill(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none" fillRule="evenodd">
-//         <path
-//           fill="currentColor"
-//           d="M13 4a1 1 0 0 1 .993.883L14 5a1 1 0 0 0 1.993.117L16 5a1 1 0 0 1 1-1h2a3 3 0 0 1 3 3v10a3 3 0 0 1-3 3h-2a1 1 0 0 1-.993-.883L16 19a1 1 0 0 0-1.993-.117L14 19a1 1 0 0 1-1 1H5a3 3 0 0 1-3-3V7a3 3 0 0 1 3-3zm2 9a1 1 0 0 0-1 1v1a1 1 0 1 0 2 0v-1a1 1 0 0 0-1-1m0-5a1 1 0 0 0-.993.883L14 9v1a1 1 0 0 0 1.993.117L16 10V9a1 1 0 0 0-1-1"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// function TimeSched(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         color="currentColor"
-//       >
-//         <path d="M12 8v4l1.5 1.5m6.045 2.953C21.182 17.337 22 17.78 22 18.5s-.818 1.163-2.455 2.047l-1.114.601c-1.257.679-1.885 1.018-2.187.772c-.74-.605.413-2.164.696-2.716c.288-.56.282-.858 0-1.408c-.283-.552-1.436-2.111-.696-2.716c.302-.246.93.093 2.187.772z"></path>
-//         <path d="M13.026 21.948Q12.52 21.999 12 22C6.477 22 2 17.523 2 12S6.477 2 12 2s10 4.477 10 10q-.002 1.03-.2 2"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function Toc(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill={props.fill}
-//         d="M3 9h14V7H3zm0 4h14v-2H3zm0 4h14v-2H3zm16 0h2v-2h-2zm0-10v2h2V7zm0 6h2v-2h-2z"
-//       ></path>
-//     </svg>
-//   );
-// }
-
-// function Tv(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1"
-//         d="M2 10c0-3.771 0-5.657 1.172-6.828S6.229 2 10 2h4c3.771 0 5.657 0 6.828 1.172S22 6.229 22 10s0 5.657-1.172 6.828S17.771 18 14 18h-4c-3.771 0-5.657 0-6.828-1.172S2 13.771 2 10m14 12a8.4 8.4 0 0 0-4-1a8.4 8.4 0 0 0-4 1"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function Undo(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         strokeWidth={0}
-//         fill="currentColor"
-//         fillRule="evenodd"
-//         d="M7.53 3.47a.75.75 0 0 1 0 1.06L5.81 6.25H15a5.75 5.75 0 0 1 0 11.5H8a.75.75 0 0 1 0-1.5h7a4.25 4.25 0 0 0 0-8.5H5.81l1.72 1.72a.75.75 0 1 1-1.06 1.06l-3-3a.75.75 0 0 1 0-1.06l3-3a.75.75 0 0 1 1.06 0"
-//         clipRule="evenodd"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function Upcoming(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1"
-//         color="currentColor"
-//       >
-//         <path d="M20.5 13.5a8.5 8.5 0 1 1-17 0a8.5 8.5 0 0 1 17 0"></path>
-//         <path d="M12 19a5.5 5.5 0 1 1 0-11m1.5-4.5a1.5 1.5 0 1 0-3 0a1.5 1.5 0 0 0 3 0m-1.5 10L15 9"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// function Upload(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M12 16.349V3.87M6.118 9.132l4.821-4.821c.293-.293.677-.44 1.061-.44m5.882 5.261l-4.821-4.821A1.5 1.5 0 0 0 12 3.87m8.75 12.645v.935a3.3 3.3 0 0 1-3.3 3.3H6.55a3.3 3.3 0 0 1-3.3-3.3v-.935"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function User(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g fill="none">
-//         <path
-//           fill="currentColor"
-//           d="M12 13c2.396 0 4.575.694 6.178 1.672c.8.488 1.484 1.064 1.978 1.69c.486.615.844 1.351.844 2.138c0 .845-.411 1.511-1.003 1.986c-.56.45-1.299.748-2.084.956c-1.578.417-3.684.558-5.913.558s-4.335-.14-5.913-.558c-.785-.208-1.524-.506-2.084-.956C3.41 20.01 3 19.345 3 18.5c0-.787.358-1.523.844-2.139c.494-.625 1.177-1.2 1.978-1.69C7.425 13.695 9.605 13 12 13m0-11a5 5 0 1 1 0 10a5 5 0 0 1 0-10"
-//         ></path>
-//       </g>
-//     </svg>
-//   );
-// }
-
-// export function UserSettings(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         d="M11.5 14.012a10.6 10.6 0 0 0-5.922 1.47c-1.415.842-5.125 2.562-2.865 4.715C3.816 21.248 5.045 22 6.59 22H12m3.5-15.5a4.5 4.5 0 1 1-9 0a4.5 4.5 0 0 1 9 0M18 20.714V22m0-1.286a3.36 3.36 0 0 1-2.774-1.43M18 20.713a3.36 3.36 0 0 0 2.774-1.43M18 14.285c1.157 0 2.176.568 2.774 1.43M18 14.287a3.36 3.36 0 0 0-2.774 1.43M18 14.287V13m4 1.929l-1.226.788M14 20.07l1.226-.788M14 14.93l1.226.788M22 20.07l-1.226-.788m0-3.566a3.12 3.12 0 0 1 0 3.566m-5.548-3.566a3.12 3.12 0 0 0 0 3.566"
-//         color="currentColor"
-//       ></path>
-//     </svg>
-//   );
-// }
-// export function UserSettings2(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <g
-//         fill="none"
-//         stroke="currentColor"
-//         strokeLinecap="round"
-//         strokeLinejoin="round"
-//         strokeWidth="1.5"
-//         color="currentColor"
-//       >
-//         <path d="m21.318 7.141l-.494-.856c-.373-.648-.56-.972-.878-1.101c-.317-.13-.676-.027-1.395.176l-1.22.344c-.459.106-.94.046-1.358-.17l-.337-.194a2 2 0 0 1-.788-.967l-.334-.998c-.22-.66-.33-.99-.591-1.178c-.261-.19-.609-.19-1.303-.19h-1.115c-.694 0-1.041 0-1.303.19c-.261.188-.37.518-.59 1.178l-.334.998a2 2 0 0 1-.789.967l-.337.195c-.418.215-.9.275-1.358.17l-1.22-.345c-.719-.203-1.078-.305-1.395-.176c-.318.129-.505.453-.878 1.1l-.493.857c-.35.608-.525.911-.491 1.234c.034.324.268.584.736 1.105l1.031 1.153c.252.319.431.875.431 1.375s-.179 1.056-.43 1.375l-1.032 1.152c-.468.521-.702.782-.736 1.105s.14.627.49 1.234l.494.857c.373.647.56.971.878 1.1s.676.028 1.395-.176l1.22-.344a2 2 0 0 1 1.359.17l.336.194c.36.23.636.57.788.968l.334.997c.22.66.33.99.591 1.18c.262.188.609.188 1.303.188h1.115c.694 0 1.042 0 1.303-.189s.371-.519.59-1.179l.335-.997c.152-.399.428-.738.788-.968l.336-.194c.42-.215.9-.276 1.36-.17l1.22.344c.718.204 1.077.306 1.394.177c.318-.13.505-.454.878-1.101l.493-.857c.35-.607.525-.91.491-1.234s-.268-.584-.736-1.105l-1.031-1.152c-.252-.32-.431-.875-.431-1.375s.179-1.056.43-1.375l1.032-1.153c.468-.52.702-.781.736-1.105s-.14-.626-.49-1.234"></path>
-//         <path d="M8.5 16a4.04 4.04 0 0 1 3.5-2.02c1.496 0 2.801.812 3.5 2.02M14 9.5a2 2 0 1 1-4 0a2 2 0 0 1 4 0"></path>
-//       </g>
-//     </svg>
-//   );
-// }
-// export function VIPIcon(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="currentColor"
-//         d="M3.005 3h18a1 1 0 0 1 1 1v16a1 1 0 0 1-1 1h-18a1 1 0 0 1-1-1V4a1 1 0 0 1 1-1m8 5.5v7h2v-7zm-.285 0H8.606l-1.497 4.113L5.612 8.5H3.498l2.611 6.964h2zm5.285 5h1.5a2.5 2.5 0 0 0 0-5h-3.5v7h2zm0-2v-1h1.5a.5.5 0 0 1 0 1z"
-//       ></path>
-//     </svg>
-//   );
-// }
-// function VIPIcon2(props: SVGProps<SVGSVGElement>) {
-//   return (
-//     <svg
-//       xmlns="http://www.w3.org/2000/svg"
-//       width="1em"
-//       height="1em"
-//       viewBox="0 0 24 24"
-//       {...props}
-//     >
-//       <path
-//         fill="currentColor"
-//         d="M2.005 19h20v2h-20zm9-11h2v8h-2zM7.97 8l-1.86 5.113L4.247 8H2.123l2.986 7.964h2L10.095 8zm9.035 6v2h-2V8h4a3 3 0 0 1 0 6zm0-4v2h2a1 1 0 1 0 0-2zm-15-7h20v2h-20z"
-//       ></path>
-//     </svg>
-//   );
-// }
-// export const icons = {
-//   Activity,
-//   AddUsers,
-//   Airplane,
-//   Airplane2,
-//   AlertTriLine,
-//   ArrowRight,
-//   ArrowRightUp,
-//   ArrowRightDouble,
-//   ArrowReturnUp,
-//   ArrowVertical,
-//   Bell,
-//   Bookmark,
-//   Bookmark2,
-//   BookmarkPlus,
-//   BookmarkCheck,
-//   BouncyArrow,
-//   BurgerRight,
-//   Calendar,
-//   CarDashboard,
-//   Chats,
-//   ChartIcon,
-//   Check,
-//   CheckDouble,
-//   ChevronDown,
-//   CloudUpload,
-//   Close,
-//   CloseLight,
-//   Cog,
-//   Cocktail,
-//   CollectionIcon,
-//   Copy,
-//   CryptoCoinbase,
-//   CryptoUSDC,
-//   DotSm,
-//   ImageIcon,
-//   ImageGen,
-//   ImageUpload,
-//   InfoLine,
-//   EGuitarFender,
-//   EGuitarGray,
-//   Energy,
-//   EventStar,
-//   EyeOpen,
-//   Fire,
-//   Fullscreen,
-//   Globe,
-//   Heart,
-//   HeartFill,
-//   HeartCheck,
-//   HotelRoom,
-//   Info,
-//   Launch,
-//   LeftChev,
-//   Lock,
-//   MailSend,
-//   MapPin,
-//   MapPin2,
-//   Minus,
-//   MoreHorizontal,
-//   MoreVertical,
-//   Mul,
-//   NumberList,
-//   PayMongo,
-//   Pen,
-//   PencilEdit,
-//   Play,
-//   Plus,
-//   PlusMinus,
-//   Printer,
-//   Printer2,
-//   QrCode,
-//   RightChev,
-//   RightLine,
-//   RightLineSm,
-//   Screen,
-//   Search,
-//   Settings,
-//   Settings2,
-//   Share,
-//   ShieldCheck,
-//   SidebarLeft,
-//   Sparkle,
-//   Sparkles,
-//   Sparkles2,
-//   SpinnerBall,
-//   SpinnerClock,
-//   SpinnersPulseRing,
-//   SpinnerDotScale,
-//   SpinnerDotMove,
-//   Stats,
-//   Squircle,
-//   Support,
-//   Tablet,
-//   Ticket,
-//   Ticket2,
-//   TicketFill,
-//   TimeSched,
-//   Toc,
-//   Tv,
-//   Undo,
-//   Upcoming,
-//   Upload,
-//   User,
-//   UserSettings,
-//   UserSettings2,
-//   VIPIcon,
-//   VIPIcon2,
-// };
