+++ src/ui/button/button-icon.tsx
@@ -1,6 +1,6 @@
 import type { ClassName } from "@/app/types";
 import type { IconName } from "@/icons/types";
-import { Iconx } from "@/icons/icon";
+import { Iconx } from "@/icons";
 import { cn } from "@/lib/utils";
 import type { ButtonHTMLAttributes } from "react";
