+++ src/ui/card/display.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Iconx } from "@/icons/icon";
+import { Iconx } from "@/icons";
 import { cn } from "@/lib/utils";
 
 interface DisplayCardProps {
